+++ src/unexploitable_search/train/mitigation_strats/train.py
@@ -28,7 +28,6 @@ from unexploitable_search.train.mitigation_strats.strats.side_obj import (
 from unexploitable_search.train.trainers import (
     GRPOTrainer,
     GRPOGameTrainer,
-    GRPOHardGameTrainer,
 )
 from unexploitable_search.train.trainers.kl_from_base_trainer import (
     KLFromBaseGRPOTrainer,
@@ -312,7 +311,8 @@ def train_with_mitigation_strat(cfg: DictConfig):
             eval_dataset=validation_dataset,  # type: ignore
             callbacks=callbacks,  # type: ignore
         )
-    elif HydraConfig.get().runtime.choices.get("strat") == "game":
+    elif HydraConfig.get().runtime.choices.get("strat") in ["game", "hard_game"]:
+        hard_mode = HydraConfig.get().runtime.choices.get("strat") == "hard_game"
         trainer = GRPOGameTrainer(
             model=model_organism,  # type: ignore
             processing_class=tokenizer,  # type: ignore
@@ -325,21 +325,7 @@ def train_with_mitigation_strat(cfg: DictConfig):
             solution_preprocessor=setting.preprocess_solution,
             preprocess_reference_solutions=cfg.strat.preprocess_reference_solutions,
             train_adversary=cfg.strat.train_adversary,
-            callbacks=callbacks,  # type: ignore
-        )
-    elif HydraConfig.get().runtime.choices.get("strat") == "hard_game":
-        trainer = GRPOHardGameTrainer(
-            model=model_organism,  # type: ignore
-            processing_class=tokenizer,  # type: ignore
-            reward_funcs=reward_funcs,  # type: ignore
-            args=args,  # type: ignore
-            train_dataset=train_dataset,  # type: ignore
-            eval_dataset=validation_dataset,  # type: ignore
-            game_coefficient=cfg.strat.two_player_coefficient,
-            num_references=cfg.strat.num_references,
-            solution_preprocessor=setting.preprocess_solution,
-            preprocess_reference_solutions=cfg.strat.preprocess_reference_solutions,
-            train_adversary=cfg.strat.train_adversary,
+            hard_mode=hard_mode,
             callbacks=callbacks,  # type: ignore
         )
     elif HydraConfig.get().runtime.choices.get("strat") == "kl_from_base_penalty":
