+++ src/unexploitable_search/train/trainers/__init__.py
@@ -1,7 +1,6 @@
 from .grpo_trainer import GRPOTrainer, GRPOConfig
 from .game_trainer.game_trainer import GRPOGameTrainer
 from .maxent_trainer import MaxEntGRPOTrainer, MaxEntGRPOConfig
-from .game_trainer.game_trainer_hard import GRPOHardGameTrainer
 
 __all__ = [
     "GRPOTrainer",
@@ -9,5 +8,4 @@ __all__ = [
     "GRPOGameTrainer",
     "MaxEntGRPOTrainer",
     "MaxEntGRPOConfig",
-    "GRPOHardGameTrainer",
 ]
