+++ src/unexploitable_search/train/trainers/game_trainer/game_trainer.py
@@ -155,11 +155,21 @@ class GRPOGameTrainer(BaseTrainer):
     A GRPOTrainer specialized for the two-player game mitigation strategy.
 
     This trainer enforces that the model must be a PEFT model and provides
-    configurable prompts for Bob and the property checker.
-    During initialization, it creates two additional LoRA adapters using
+    configurable prompts for Bob/Charlie and the property checker.
+    During initialization, it creates additional LoRA adapters using
     the same config as existing adapters:
     - 'property_checker': For property verification (not trained)
-    - 'bob': For Bob's completions (will be trained)
+    - 'charlie': For Charlie's completions (will be trained)
+    - 'bob': For Bob's completions (will be trained, only in hard mode)
+
+    In regular mode (hard=False, default):
+    - Charlie generates both alternative solutions and properties
+    - Only one adversary adapter (charlie) is created
+
+    In hard mode (hard=True):
+    - Bob generates properties only
+    - Charlie generates alternative solutions only
+    - Two adversary adapters (bob and charlie) are created
 
     Example:
 
@@ -297,6 +307,7 @@ class GRPOGameTrainer(BaseTrainer):
         ] = (None, None),
         peft_config: Optional["PeftConfig"] = None,
         charlie_prompt: Optional[str] = None,
+        bob_prompt: Optional[str] = None,
         property_checker_prompt: Optional[str] = None,
         game_coefficient: float = 0.2,
         tolerance: float = 0.1,
@@ -306,6 +317,7 @@ class GRPOGameTrainer(BaseTrainer):
         num_references: int = 1,
         game_score_weights: list[float] | None = None,
         train_adversary: bool = True,
+        hard_mode: bool = False,
     ):
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         has_preprocessor = solution_preprocessor is not None
@@ -332,12 +344,20 @@ class GRPOGameTrainer(BaseTrainer):
                 "passing it to this trainer."
             )
 
+        # Store hard mode flag
+        self.hard_mode = hard_mode
+
         # Set prompts with defaults
         self.charlie_prompt = (
             charlie_prompt
             if charlie_prompt is not None
             else self.DEFAULT_CHARLIE_PROMPT
         )
+        # Bob prompt is only used in hard mode
+        if self.hard_mode:
+            self.bob_prompt = (
+                bob_prompt if bob_prompt is not None else self.DEFAULT_BOB_PROMPT
+            )
         self.property_checker_prompt = (
             property_checker_prompt
             if property_checker_prompt is not None
@@ -376,15 +396,23 @@ class GRPOGameTrainer(BaseTrainer):
         model.add_adapter(self.property_checker_adapter_name, base_lora_config)
 
         # Charlie's adapter - will be used for Charlie's generations (will be trained)
-        # Charlie generates both alternative solutions and properties
         self.charlie_adapter_name = "charlie"
         model.add_adapter(self.charlie_adapter_name, base_lora_config)
 
-        # Only enable gradients for adversary adapters if training them
-        if self.train_adversary:
-            for name, param in model.named_parameters():
-                if self.charlie_adapter_name in name:
-                    param.requires_grad = True
+        trainable_adversary_adapters = [self.charlie_adapter_name]
+        if self.hard_mode:
+            # Hard mode: Bob and Charlie are separate
+            # Bob's adapter - will be used for Bob's generations (will be trained)
+            self.bob_adapter_name = "bob"
+            model.add_adapter(self.bob_adapter_name, base_lora_config)
+
+            trainable_adversary_adapters.append(self.bob_adapter_name)
+
+            # Only enable gradients for adversary adapters if training them
+            if self.train_adversary:
+                for name, param in model.named_parameters():
+                    if name in trainable_adversary_adapters:
+                        param.requires_grad = True
         ############### END GRPOGAMETRAINER CHANGE ###############
 
         # Args
@@ -859,6 +887,14 @@ class GRPOGameTrainer(BaseTrainer):
         self._charlie_completion_ids: List[List[int]] = []
         self._charlie_logprobs: Optional[List[Any]] = None
 
+        # Storage for Bob's completions (only used in hard mode, but initialize to avoid AttributeError)
+        self._bob_prompt_ids: List[List[int]] = []
+        self._bob_completion_ids: List[List[int]] = []
+        self._bob_logprobs: Optional[List[Any]] = None
+        self._bob_prompts: List[Any] = []
+        self._bob_completions: List[Any] = []
+        self._bob_properties: List[str] = []
+
         # Add game as an additional reward function
         self.reward_func_names.append("game_reward")
         self.reward_weights = torch.cat(
@@ -871,12 +907,31 @@ class GRPOGameTrainer(BaseTrainer):
         # Initialize Charlie-specific log entries
         self._logs["charlie_prompt"] = deque(maxlen=args.generation_batch_size)
         self._logs["charlie_completion"] = deque(maxlen=args.generation_batch_size)
+
+        # Initialize Bob-specific log entries (only in hard mode)
+        if self.hard_mode:
+            self._logs["bob_prompt"] = deque(maxlen=args.generation_batch_size)
+            self._logs["bob_completion"] = deque(maxlen=args.generation_batch_size)
         ############### END GRPOGAMETRAINER CHANGE ###############
 
     @property
     def reference_logprobs(self) -> Optional[List[Any]]:
         """Access to the stored reference logprobs."""
         return self._reference_logprobs
+
+    def get_bob_prompt(self, property_description: str) -> str:
+        """
+        Format Bob's prompt with a specific property description.
+
+        Args:
+            property_description: The property that Bob should avoid
+
+        Returns:
+            Formatted prompt string
+        """
+        if not self.hard_mode:
+            raise ValueError("get_bob_prompt is only available in hard mode")
+        return self.bob_prompt.format(property_description=property_description)
 
     def get_property_checker_prompt(self) -> str:
         """
@@ -886,6 +941,29 @@ class GRPOGameTrainer(BaseTrainer):
             Property checker prompt string
         """
         return self.property_checker_prompt
+
+    def _format_bob_prompts(
+        self,
+        original_prompts_batch: List[List[str]],
+        reference_solutions_batch: List[List[str]],
+    ) -> List[List[Dict[str, Any]]]:
+        """
+        Format prompts for Bob given the original prompts and reference solutions.
+
+        Bob identifies removable properties from the reference solutions.
+
+        Args:
+            original_prompts_batch: List of original prompt messages (chat format)
+            reference_solutions_batch: List of reference solution code strings
+
+        Returns:
+            List of formatted prompts for Bob (chat format)
+        """
+        if not self.hard_mode:
+            raise ValueError("_format_bob_prompts is only available in hard mode")
+        return self._format_player_prompts(
+            original_prompts_batch, reference_solutions_batch, self.bob_prompt
+        )
 
     def _format_player_prompts(
         self,
@@ -1602,8 +1680,16 @@ class GRPOGameTrainer(BaseTrainer):
                     charlie_solutions_text.append(str(solution))
 
             solutions_to_check = alice_solutions + charlie_solutions_text
-            # Use properties extracted from Charlie's completions
-            properties_to_check = self._charlie_properties * 2
+            # Use properties based on mode
+            if self.hard_mode:
+                # Hard mode: Bob generates properties
+                properties_to_check = self._bob_properties * 2
+                properties_for_game = self._bob_properties
+            else:
+                # Regular mode: Charlie generates properties
+                properties_to_check = self._charlie_properties * 2
+                properties_for_game = self._charlie_properties
+
             if self.solution_preprocessor is not None:
                 solutions_to_check = [
                     self.solution_preprocessor(solution)
@@ -1629,11 +1715,12 @@ class GRPOGameTrainer(BaseTrainer):
             )
 
             # Compute game rewards
-            # Charlie's solutions compete with Alice's, using properties extracted from Charlie
+            # In hard mode: Bob generates properties, Charlie generates solutions
+            # In regular mode: Charlie generates both properties and solutions
             self._game_rewards = self._get_game_rewards(
                 alice_solutions=alice_solutions,
                 bob_solutions=charlie_solutions_text,  # Using Charlie's solutions
-                properties=self._charlie_properties,  # Using properties from Charlie
+                properties=properties_for_game,
                 has_properties_alice=has_properties_alice,
                 has_properties_bob=has_properties_charlie,  # Charlie's property flags
                 alice_scores=alice_scores,
@@ -1641,12 +1728,17 @@ class GRPOGameTrainer(BaseTrainer):
             )
 
             if is_wandb_available():
+                # Always use the same columns regardless of mode
+                # Some fields will be None when not applicable
                 table = wandb.Table(
                     columns=[
                         "question",
                         "alice_completion",
+                        "bob_completion",
                         "charlie_completion",
-                        "property",
+                        "bob_property",
+                        "charlie_property",
+                        "property",  # Bob's property in hard mode, Charlie's in regular mode
                         "alice_flagged",
                         "charlie_solution",
                         "charlie_flagged",
@@ -1656,11 +1748,28 @@ class GRPOGameTrainer(BaseTrainer):
                     ]
                 )
                 for i in range(len(prompts) // 2):
+                    # Determine values based on mode
+                    # Property: Bob's in hard mode, Charlie's in regular mode
+                    bob_property = None
+                    bob_completion = None
+                    property_value = self._charlie_properties[
+                        i
+                    ]  # Charlie's property in regular mode
+                    charlie_property = self._charlie_properties[i]
+                    charlie_completion = self._charlie_completions[i][0]["content"]
+                    if self.hard_mode:
+                        bob_property = self._bob_properties[i]
+                        bob_completion = self._bob_completions[i][0]["content"]
+                        property_value = self._bob_properties[i]
+
                     table.add_data(
                         prompts[i][1]["content"],
                         alice_solutions[i],
-                        self._charlie_completions[i][0]["content"],
-                        self._charlie_properties[i],  # Property from Charlie
+                        bob_completion,
+                        charlie_completion,
+                        bob_property,
+                        charlie_property,
+                        property_value,
                         has_properties_alice[i],
                         charlie_solutions_text[i],
                         has_properties_charlie[i],
@@ -1694,6 +1803,13 @@ class GRPOGameTrainer(BaseTrainer):
         We temporarily increase self.num_generations to n+1, call parent's
         generation, then restore it. The reference completion
         (last one in each group) will be extracted and stored.
+
+        Always returns a 12-element tuple (hard mode format):
+        (prompt_ids, completion_ids, logprobs, forward_kwargs,
+         bob_prompt_ids, bob_completion_ids, bob_logprobs, bob_forward_kwargs,
+         charlie_prompt_ids, charlie_completion_ids, charlie_logprobs, charlie_forward_kwargs)
+
+        When not in hard mode, bob_* values are None.
         """
         # During evaluation, skip all game mechanisms and use core implementation
         if not self.model.training:
@@ -1701,72 +1817,187 @@ class GRPOGameTrainer(BaseTrainer):
                 self._generate_single_turn_core(prompts, images)
             )
 
+            # Always return the full tuple structure (hard mode format)
+            # Bob values are None when not in hard mode
             return (
                 prompt_ids,
                 completion_ids,
                 logprobs,
                 forward_kwargs,
-                None,
-                None,
-                None,
-                None,
+                None,  # bob_prompt_ids
+                None,  # bob_completion_ids
+                None,  # bob_logprobs
+                None,  # bob_forward_kwargs
+                None,  # charlie_prompt_ids
+                None,  # charlie_completion_ids
+                None,  # charlie_logprobs
+                None,  # charlie_forward_kwargs
             )
 
         # Store original num_generations
         original_num_gens = self.num_generations
 
         try:
-            # Temporarily set to n + num_references for generation
-            # Charlie will see reference solutions
-            self.num_generations = original_num_gens + self.num_references
-            prompt_ids, completion_ids, logprobs, forward_kwargs = (
-                self._generate_single_turn_core(prompts, images)
-            )
+            if self.hard_mode:
+                # Hard mode: Bob and Charlie each need num_references, so we need 2 * num_references total
+                # Temporarily set to n + 2*num_references for generation
+                # Bob and Charlie will each see different reference solutions
+                self.num_generations = original_num_gens + 2 * self.num_references
+                prompt_ids, completion_ids, logprobs, forward_kwargs = (
+                    self._generate_single_turn_core(prompts, images)
+                )
 
-            # Extract reference completions into list of lists
-            # Structure: batch_size * (original_num_gens + num_references) total completions
-            # Each batch item has:
-            #   - original_num_gens regular completions (for Alice)
-            #   - num_references reference completions (for Charlie)
-            charlie_reference_completion_ids = [
-                completion_ids[
-                    batch_start + original_num_gens : batch_start
-                    + original_num_gens
-                    + self.num_references
+                # Extract reference completions into list of lists
+                # Structure: batch_size * (original_num_gens + 2*num_references) total completions
+                # Each batch item has:
+                #   - original_num_gens regular completions (for Alice)
+                #   - num_references reference completions (for Bob)
+                #   - num_references reference completions (for Charlie)
+                bob_reference_completion_ids = [
+                    completion_ids[
+                        batch_start + original_num_gens : batch_start
+                        + original_num_gens
+                        + self.num_references
+                    ]
+                    for batch_start in range(
+                        0, len(completion_ids), self.num_generations
+                    )
                 ]
-                for batch_start in range(0, len(completion_ids), self.num_generations)
-            ]
-            charlie_reference_prompts = [
-                prompts[
-                    batch_start + original_num_gens : batch_start
-                    + original_num_gens
-                    + self.num_references
+                bob_reference_prompts = [
+                    prompts[
+                        batch_start + original_num_gens : batch_start
+                        + original_num_gens
+                        + self.num_references
+                    ]
+                    for batch_start in range(
+                        0, len(completion_ids), self.num_generations
+                    )
                 ]
-                for batch_start in range(0, len(completion_ids), self.num_generations)
-            ]
 
-            self.num_generations = original_num_gens
+                charlie_reference_completion_ids = [
+                    completion_ids[
+                        batch_start
+                        + original_num_gens
+                        + self.num_references : batch_start
+                        + original_num_gens
+                        + 2 * self.num_references
+                    ]
+                    for batch_start in range(
+                        0, len(completion_ids), self.num_generations
+                    )
+                ]
+                charlie_reference_prompts = [
+                    prompts[
+                        batch_start
+                        + original_num_gens
+                        + self.num_references : batch_start
+                        + original_num_gens
+                        + 2 * self.num_references
+                    ]
+                    for batch_start in range(
+                        0, len(completion_ids), self.num_generations
+                    )
+                ]
 
-            # Generate Charlie's completions (alternative solutions and properties) using Charlie's references
-            (
-                charlie_prompt_ids,
-                charlie_completion_ids,
-                charlie_logprobs,
-                charlie_forward_kwargs,
-            ) = self._generate_charlie_completions(
-                charlie_reference_completion_ids, charlie_reference_prompts, images
-            )
+                self.num_generations = original_num_gens
 
-            return (
-                prompt_ids,
-                completion_ids,
-                logprobs,
-                forward_kwargs,
-                charlie_prompt_ids,
-                charlie_completion_ids,
-                charlie_logprobs,
-                charlie_forward_kwargs,
-            )
+                # Generate Bob's completions (properties only) using Bob's references
+                (
+                    bob_prompt_ids,
+                    bob_completion_ids,
+                    bob_logprobs,
+                    bob_forward_kwargs,
+                ) = self._generate_bob_completions(
+                    bob_reference_completion_ids, bob_reference_prompts, images
+                )
+
+                # Generate Charlie's completions (alternative solutions only) using Charlie's references
+                (
+                    charlie_prompt_ids,
+                    charlie_completion_ids,
+                    charlie_logprobs,
+                    charlie_forward_kwargs,
+                ) = self._generate_charlie_completions(
+                    charlie_reference_completion_ids, charlie_reference_prompts, images
+                )
+
+                return (
+                    prompt_ids,
+                    completion_ids,
+                    logprobs,
+                    forward_kwargs,
+                    bob_prompt_ids,
+                    bob_completion_ids,
+                    bob_logprobs,
+                    bob_forward_kwargs,
+                    charlie_prompt_ids,
+                    charlie_completion_ids,
+                    charlie_logprobs,
+                    charlie_forward_kwargs,
+                )
+            else:
+                # Regular mode: Charlie generates both solutions and properties
+                # Temporarily set to n + num_references for generation
+                # Charlie will see reference solutions
+                self.num_generations = original_num_gens + self.num_references
+                prompt_ids, completion_ids, logprobs, forward_kwargs = (
+                    self._generate_single_turn_core(prompts, images)
+                )
+
+                # Extract reference completions into list of lists
+                # Structure: batch_size * (original_num_gens + num_references) total completions
+                # Each batch item has:
+                #   - original_num_gens regular completions (for Alice)
+                #   - num_references reference completions (for Charlie)
+                charlie_reference_completion_ids = [
+                    completion_ids[
+                        batch_start + original_num_gens : batch_start
+                        + original_num_gens
+                        + self.num_references
+                    ]
+                    for batch_start in range(
+                        0, len(completion_ids), self.num_generations
+                    )
+                ]
+                charlie_reference_prompts = [
+                    prompts[
+                        batch_start + original_num_gens : batch_start
+                        + original_num_gens
+                        + self.num_references
+                    ]
+                    for batch_start in range(
+                        0, len(completion_ids), self.num_generations
+                    )
+                ]
+
+                self.num_generations = original_num_gens
+
+                # Generate Charlie's completions (alternative solutions and properties) using Charlie's references
+                (
+                    charlie_prompt_ids,
+                    charlie_completion_ids,
+                    charlie_logprobs,
+                    charlie_forward_kwargs,
+                ) = self._generate_charlie_completions(
+                    charlie_reference_completion_ids, charlie_reference_prompts, images
+                )
+
+                # Always return the full tuple structure (hard mode format)
+                # Bob values are None in regular mode
+                return (
+                    prompt_ids,
+                    completion_ids,
+                    logprobs,
+                    forward_kwargs,
+                    None,  # bob_prompt_ids (not used in regular mode)
+                    None,  # bob_completion_ids (not used in regular mode)
+                    None,  # bob_logprobs (not used in regular mode)
+                    None,  # bob_forward_kwargs (not used in regular mode)
+                    charlie_prompt_ids,
+                    charlie_completion_ids,
+                    charlie_logprobs,
+                    charlie_forward_kwargs,
+                )
         finally:
             self.num_generations = original_num_gens
         ############### END GRPOGAMETRAINER CHANGE ###############
@@ -2129,14 +2360,30 @@ class GRPOGameTrainer(BaseTrainer):
         return prompt_ids, completion_ids, logprobs, forward_kwargs
 
     def _generate(self, prompts: list[str], images: Optional[list]):
+        """
+        Generate completions for prompts, including adversary completions.
+
+        Always returns a 16-element tuple (hard mode format):
+        (prompt_ids, completion_ids, total_completion_tokens, logprobs, forward_kwargs,
+         bob_prompt_ids, bob_completion_ids, bob_total_completion_tokens, bob_logprobs, bob_forward_kwargs,
+         charlie_prompt_ids, charlie_completion_ids, charlie_total_completion_tokens, charlie_logprobs, charlie_forward_kwargs)
+
+        When not in hard mode, bob_* values are None (except bob_total_completion_tokens which is 0).
+        """
         device = self.accelerator.device
         mode = "train" if self.model.training else "eval"
 
+        # Always unpack the full tuple structure (hard mode format)
+        # Bob values will be None when not in hard mode
         (
             prompt_ids,
             completion_ids,
             logprobs,
             forward_kwargs,
+            bob_prompt_ids,
+            bob_completion_ids,
+            bob_logprobs,
+            bob_forward_kwargs,
             charlie_prompt_ids,
             charlie_completion_ids,
             charlie_logprobs,
@@ -2208,20 +2455,122 @@ class GRPOGameTrainer(BaseTrainer):
             _, charlie_total_completion_tokens = log_lengths(
                 charlie_prompt_ids, charlie_completion_ids, "charlie"
             )
+
+        bob_total_completion_tokens = 0
+        if self.hard_mode and bob_prompt_ids is not None:
+            _, bob_total_completion_tokens = log_lengths(
+                bob_prompt_ids, bob_completion_ids, "bob"
+            )
         ############### END GRPOGAMETRAINER CHANGE ###############
 
+        # Always return the full tuple structure (hard mode format)
+        # Bob values will be None when not in hard mode
         return (
             prompt_ids,
             completion_ids,
             total_completion_tokens,
             logprobs,
             forward_kwargs,
+            bob_prompt_ids,  # None when not in hard mode
+            bob_completion_ids,  # None when not in hard mode
+            bob_total_completion_tokens,  # 0 when not in hard mode
+            bob_logprobs,  # None when not in hard mode
+            bob_forward_kwargs,  # None when not in hard mode
             charlie_prompt_ids,
             charlie_completion_ids,
             charlie_total_completion_tokens,
             charlie_logprobs,
             charlie_forward_kwargs,
         )
+
+    def _generate_bob_completions(
+        self,
+        reference_completion_ids: List[List[List[int]]],
+        reference_prompts: List[str],
+        images: Optional[list],
+    ):
+        """
+        Generate Bob's completions (properties only).
+
+        Bob receives problem + reference solutions and generates only properties.
+
+        Args:
+            reference_completion_ids: Reference solution token IDs
+            reference_prompts: Original problem prompts
+            images: Optional images
+
+        Returns:
+            Tuple of (prompt_ids, completion_ids, logprobs, forward_kwargs)
+        """
+        if not self.hard_mode:
+            raise ValueError("_generate_bob_completions is only available in hard mode")
+
+        current_active_adapter = self.model.active_adapter
+
+        self.model.set_adapter(self.bob_adapter_name)
+        try:
+            # Flatten reference_completion_ids from List[List[List[int]]] to List[List[int]]
+            # Structure: batch_size * num_references -> flat list
+            batch_size = len(reference_completion_ids)
+            num_references = len(reference_completion_ids[0]) if batch_size > 0 else 0
+            flattened_completion_ids = [
+                completion_id
+                for batch_item in reference_completion_ids
+                for completion_id in batch_item
+            ]
+
+            # Decode all at once
+            reference_solutions = self.processing_class.batch_decode(
+                flattened_completion_ids, skip_special_tokens=True
+            )
+
+            # Preprocess if needed
+            if (
+                self.preprocess_reference_solutions
+                and self.solution_preprocessor is not None
+            ):
+                reference_solutions = [
+                    self.solution_preprocessor(solution)
+                    for solution in reference_solutions
+                ]
+
+            # Reshape back to batch_size * num_references structure
+            reference_solutions = [
+                reference_solutions[i * num_references : (i + 1) * num_references]
+                for i in range(batch_size)
+            ]
+
+            prompts = self._format_bob_prompts(reference_prompts, reference_solutions)
+            # Call parent's generation
+            prompt_ids, completion_ids, logprobs, forward_kwargs = (
+                self._generate_single_turn_core(prompts, images)
+            )
+
+            # Decode
+            completions_text = self.processing_class.batch_decode(
+                completion_ids, skip_special_tokens=True
+            )
+            completions = []
+            for prompt, completion in zip(prompts, completions_text):
+                # Access bootstrap without mutating the prompt list
+                bootstrap = (
+                    prompt[-1]["content"] if prompt[-1]["role"] == "assistant" else ""
+                )
+                completions.append(
+                    [{"role": "assistant", "content": bootstrap + completion}]
+                )
+
+        finally:
+            self.model.set_adapter(current_active_adapter)
+
+        self._bob_prompts = prompts
+        self._bob_completions = completions
+        self._bob_completion_ids = completion_ids
+
+        # Bob now only generates properties, not solutions
+        self._bob_properties = self._extract_properties(completions)
+
+        return (prompt_ids, completion_ids, logprobs, forward_kwargs)
 
     def _generate_charlie_completions(
         self,
@@ -2230,10 +2579,10 @@ class GRPOGameTrainer(BaseTrainer):
         images: Optional[list],
     ):
         """
-        Generate Charlie's completions (alternative solutions and properties).
+        Generate Charlie's completions.
 
-        Charlie receives problem + reference solutions and generates both
-        alternative solutions and properties extracted from those solutions.
+        In regular mode: Charlie generates both alternative solutions and properties.
+        In hard mode: Charlie generates only alternative solutions (Bob generates properties).
 
         Args:
             reference_completion_ids: Reference solution token IDs
@@ -2838,13 +3187,19 @@ class GRPOGameTrainer(BaseTrainer):
         else:
             images = None
 
+        # Always unpack the full tuple structure (hard mode format)
+        # Bob values will be None when not in hard mode
         (
             prompt_ids_list,
             completion_ids_list,
             num_items_in_batch,
             sampling_per_token_logps_list,
             forward_kwargs,
-            # Assume for evals these are None
+            bob_prompt_ids_list,
+            bob_completion_ids_list,
+            bob_num_items_in_batch,
+            bob_sampling_per_token_logps_list,
+            bob_forward_kwargs,
             charlie_prompt_ids_list,
             charlie_completion_ids_list,
             charlie_num_items_in_batch,
@@ -2855,8 +3210,12 @@ class GRPOGameTrainer(BaseTrainer):
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         if self.model.training:
             # Filter out the reference completions from the training completions
-            # Now we have num_references (for Charlie)
-            total_per_batch = self.num_generations + self.num_references
+            if self.hard_mode:
+                # Hard mode: Bob and Charlie each need num_references, so we need 2 * num_references total
+                total_per_batch = self.num_generations + 2 * self.num_references
+            else:
+                # Regular mode: Charlie needs num_references
+                total_per_batch = self.num_generations + self.num_references
             train_mask = [
                 i % total_per_batch < self.num_generations for i in range(len(prompts))
             ]
@@ -2903,10 +3262,44 @@ class GRPOGameTrainer(BaseTrainer):
 
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         if self.model.training:
-            # Prepare Charlie's inputs
+            original_adapter = self.model.active_adapter
+
+            if self.hard_mode:
+                # Hard mode: Prepare Bob's inputs
+                bob_prompts = self._bob_prompts
+                bob_inputs = [{"prompt": prompt} for prompt in bob_prompts]
+                # Set requires grad to True for Bob's adapter
+                self.model.set_adapter(self.bob_adapter_name)
+                try:
+                    (
+                        bob_prompt_ids,
+                        bob_prompt_mask,
+                        bob_completion_ids,
+                        bob_completion_mask,
+                        bob_completions,
+                        bob_prompts_text,
+                        bob_completions_text,
+                        bob_old_per_token_logps,
+                        bob_ref_per_token_logps,
+                        bob_sampling_per_token_logps,
+                        bob_importance_sampling_ratio,
+                        bob_num_images,
+                    ) = self._prepare_scoring_inputs(
+                        bob_inputs,
+                        bob_prompts,
+                        images,
+                        bob_prompt_ids_list,
+                        bob_completion_ids_list,
+                        bob_sampling_per_token_logps_list,
+                        bob_forward_kwargs,
+                    )
+                finally:
+                    # Restore the original adapter
+                    self.model.set_adapter(original_adapter)
+
+            # Prepare Charlie's inputs similarly
             charlie_prompts = self._charlie_prompts
             charlie_inputs = [{"prompt": prompt} for prompt in charlie_prompts]
-            original_adapter = self.model.active_adapter
             self.model.set_adapter(self.charlie_adapter_name)
             try:
                 (
@@ -2954,6 +3347,7 @@ class GRPOGameTrainer(BaseTrainer):
         # We compute charlie's advantages from negated game rewards (only during training)
         # Charlie gets negative game rewards (adversary)
         # All other rewards are zeroed out for Charlie
+        # In hard mode, Bob uses the same advantages as Charlie (same team)
         if self.model.training:
             charlie_rewards_per_func = rewards_per_func.clone()
             charlie_rewards_per_func[:, :-1] = 0
@@ -2961,6 +3355,10 @@ class GRPOGameTrainer(BaseTrainer):
             charlie_advantages, *_ = self._calculate_advantages(
                 prompts, images, charlie_rewards_per_func
             )
+
+            # In hard mode, Bob uses the same advantages as Charlie (same team)
+            if self.hard_mode:
+                bob_advantages = charlie_advantages.clone()
         ############### END GRPOGAMETRAINER CHANGE ###############
 
         # Calculate mean reward per function, but only for samples where the function was applied (non-NaN values)
@@ -2984,6 +3382,9 @@ class GRPOGameTrainer(BaseTrainer):
         self._logs["completion"].extend(gather_object(completions_text))
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         if self.model.training:
+            if self.hard_mode:
+                self._logs["bob_prompt"].extend(gather_object(bob_prompts_text))
+                self._logs["bob_completion"].extend(gather_object(bob_completions_text))
             self._logs["charlie_prompt"].extend(gather_object(charlie_prompts_text))
             self._logs["charlie_completion"].extend(
                 gather_object(charlie_completions_text)
@@ -3054,6 +3455,17 @@ class GRPOGameTrainer(BaseTrainer):
         }
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         if self.model.training:
+            if self.hard_mode:
+                output.update(
+                    {
+                        "bob_prompt_ids": bob_prompt_ids,
+                        "bob_prompt_mask": bob_prompt_mask,
+                        "bob_completion_ids": bob_completion_ids,
+                        "bob_completion_mask": bob_completion_mask,
+                        "bob_advantages": bob_advantages,
+                        "bob_num_items_in_batch": bob_num_items_in_batch,
+                    }
+                )
             output.update(
                 {
                     "charlie_prompt_ids": charlie_prompt_ids,
@@ -3069,6 +3481,9 @@ class GRPOGameTrainer(BaseTrainer):
             output["old_per_token_logps"] = old_per_token_logps
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         if self.model.training:
+            if self.hard_mode:
+                if bob_old_per_token_logps is not None:
+                    output["bob_old_per_token_logps"] = bob_old_per_token_logps
             if charlie_old_per_token_logps is not None:
                 output["charlie_old_per_token_logps"] = charlie_old_per_token_logps
         ############### END GRPOGAMETRAINER CHANGE ###############
@@ -3076,6 +3491,10 @@ class GRPOGameTrainer(BaseTrainer):
             output["importance_sampling_ratio"] = importance_sampling_ratio
             ############### BEGIN GRPOGAMETRAINER CHANGE ###############
             if self.model.training:
+                if self.hard_mode:
+                    output["bob_importance_sampling_ratio"] = (
+                        bob_importance_sampling_ratio
+                    )
                 output["charlie_importance_sampling_ratio"] = (
                     charlie_importance_sampling_ratio
                 )
@@ -3084,6 +3503,9 @@ class GRPOGameTrainer(BaseTrainer):
             output["ref_per_token_logps"] = ref_per_token_logps
         ############### BEGIN GRPOGAMETRAINER CHANGE ###############
         if self.model.training:
+            if self.hard_mode:
+                if bob_ref_per_token_logps is not None:
+                    output["bob_ref_per_token_logps"] = bob_ref_per_token_logps
             if charlie_ref_per_token_logps is not None:
                 output["charlie_ref_per_token_logps"] = charlie_ref_per_token_logps
         ############### END GRPOGAMETRAINER CHANGE ###############
@@ -3157,15 +3579,22 @@ class GRPOGameTrainer(BaseTrainer):
         self, model, inputs, return_outputs=False, num_items_in_batch=None
     ):
         """
-        Override compute_loss to train Alice and Charlie simultaneously.
+        Override compute_loss to train Alice and Charlie simultaneously (or Alice, Bob, and Charlie in hard mode).
 
-        This method:
+        Regular mode:
         1. Computes Alice's loss with existing game reward advantages
         2. Switches to Charlie's adapter and computes Charlie's loss (alternative solutions and properties)
         3. Charlie uses negative game rewards (adversary)
-        4. Returns both losses
+        4. Returns (alice_loss, None, charlie_loss) - bob_loss is None in regular mode
 
-        When train_adversary=False, only Alice's loss is computed and Charlie loss is None.
+        Hard mode:
+        1. Computes Alice's loss with existing game reward advantages
+        2. Switches to Bob's adapter and computes Bob's loss (property identification)
+        3. Switches to Charlie's adapter and computes Charlie's loss (alternative solutions)
+        4. Bob and Charlie both use negative game rewards (same team)
+        5. Returns (alice_loss, bob_loss, charlie_loss)
+
+        When train_adversary=False, only Alice's loss is computed and adversary losses are None.
 
         Args:
             model: The model being trained
@@ -3174,7 +3603,9 @@ class GRPOGameTrainer(BaseTrainer):
             (not supported for combined loss)
             num_items_in_batch: Number of items in the batch
         Returns:
-            Tuple of (alice_loss, charlie_loss) where charlie_loss is None if train_adversary=False
+            Always returns a 3-element tuple: (alice_loss, bob_loss, charlie_loss)
+            - bob_loss is None when not in hard mode or when train_adversary=False
+            - charlie_loss is None when train_adversary=False
         """
         # During evaluation, skip all game mechanisms and use parent's implementation
         if not self.model.training:
@@ -3190,28 +3621,71 @@ class GRPOGameTrainer(BaseTrainer):
         # in _calculate_rewards)
         alice_loss = self.compute_loss_core(model, inputs, return_outputs=False)
 
-        # Skip Charlie loss computation if train_adversary is disabled
+        # Skip adversary loss computation if train_adversary is disabled
         if not self.train_adversary:
-            return alice_loss, None
+            # Always return 3 elements: (alice_loss, bob_loss, charlie_loss)
+            # Bob and Charlie losses are None when train_adversary=False
+            return alice_loss, None, None
 
-        # Switch to Charlie's adapter to train alternative solution generation and property extraction
-        self.model.set_adapter(self.charlie_adapter_name)
-        try:
-            charlie_inputs = inputs.copy()
-            # Construct Charlie's inputs
-            for key, value in inputs.items():
-                if key.startswith("charlie_"):
-                    charlie_inputs[key.replace("charlie_", "")] = value
+        if self.hard_mode:
+            # Hard mode: Train Bob and Charlie separately
+            # Switch to Bob's adapter to train property identification
+            self.model.set_adapter(self.bob_adapter_name)
+            try:
+                bob_inputs = inputs.copy()
+                # Construct Bob's inputs
+                for key, value in inputs.items():
+                    if key.startswith("bob_"):
+                        bob_inputs[key.replace("bob_", "")] = value
 
-            # Compute Charlie's loss
-            charlie_loss = self.compute_loss_core(
-                model, charlie_inputs, return_outputs=False
-            )
-        finally:
-            # Restore the original adapter
-            self.model.set_adapter(original_adapter)
+                # Compute Bob's loss
+                bob_loss = self.compute_loss_core(
+                    model, bob_inputs, return_outputs=False
+                )
+            finally:
+                # Restore the original adapter
+                self.model.set_adapter(original_adapter)
 
-        return alice_loss, charlie_loss
+            # Switch to Charlie's adapter to train alternative solution generation
+            self.model.set_adapter(self.charlie_adapter_name)
+            try:
+                charlie_inputs = inputs.copy()
+                # Construct Charlie's inputs (use Bob's advantages since they're on the same team)
+                for key, value in inputs.items():
+                    if key.startswith("charlie_"):
+                        charlie_inputs[key.replace("charlie_", "")] = value
+
+                # Compute Charlie's loss
+                charlie_loss = self.compute_loss_core(
+                    model, charlie_inputs, return_outputs=False
+                )
+            finally:
+                # Restore the original adapter
+                self.model.set_adapter(original_adapter)
+
+            return alice_loss, bob_loss, charlie_loss
+        else:
+            # Regular mode: Train only Charlie (who generates both solutions and properties)
+            # Switch to Charlie's adapter to train alternative solution generation and property extraction
+            self.model.set_adapter(self.charlie_adapter_name)
+            try:
+                charlie_inputs = inputs.copy()
+                # Construct Charlie's inputs
+                for key, value in inputs.items():
+                    if key.startswith("charlie_"):
+                        charlie_inputs[key.replace("charlie_", "")] = value
+
+                # Compute Charlie's loss
+                charlie_loss = self.compute_loss_core(
+                    model, charlie_inputs, return_outputs=False
+                )
+            finally:
+                # Restore the original adapter
+                self.model.set_adapter(original_adapter)
+
+            # Always return 3 elements: (alice_loss, bob_loss, charlie_loss)
+            # Bob loss is None in regular mode
+            return alice_loss, None, charlie_loss
 
     def compute_loss_core(
         self, model, inputs, return_outputs=False, num_items_in_batch=None
@@ -3438,7 +3912,9 @@ class GRPOGameTrainer(BaseTrainer):
                 return loss_mb.reduce_mean().detach().to(self.args.device)
 
             with self.compute_loss_context_manager():
-                alice_loss, charlie_loss = self.compute_loss(
+                # Always unpack 3 elements: (alice_loss, bob_loss, charlie_loss)
+                # bob_loss is None when not in hard mode or when train_adversary=False
+                alice_loss, bob_loss, charlie_loss = self.compute_loss(
                     model, inputs, num_items_in_batch=num_items_in_batch
                 )
 
@@ -3510,10 +3986,17 @@ class GRPOGameTrainer(BaseTrainer):
             self.model.set_adapter(original_adapter)
             alice_loss = loss_backward(alice_loss)
 
-            # Only compute backward pass for Charlie if train_adversary is enabled
-            if self.train_adversary and charlie_loss is not None:
-                self.model.set_adapter(self.charlie_adapter_name)
-                charlie_loss = loss_backward(charlie_loss)
+            # Only compute backward pass for adversary losses if train_adversary is enabled
+            if self.train_adversary:
+                if self.hard_mode and bob_loss is not None:
+                    # Hard mode: Backward pass for Bob
+                    self.model.set_adapter(self.bob_adapter_name)
+                    bob_loss = loss_backward(bob_loss)
+
+                if charlie_loss is not None:
+                    # Backward pass for Charlie
+                    self.model.set_adapter(self.charlie_adapter_name)
+                    charlie_loss = loss_backward(charlie_loss)
 
             self.model.set_adapter(original_adapter)
 
@@ -3529,8 +4012,9 @@ class GRPOGameTrainer(BaseTrainer):
         inputs = self._prepare_inputs(inputs)
         with torch.no_grad():
             with self.compute_loss_context_manager():
-                loss = self.compute_loss(model, inputs)
-            loss = loss.mean().detach()
+                # Always unpack 3 elements, but only use alice_loss for prediction step
+                alice_loss, _, _ = self.compute_loss(model, inputs)
+            loss = alice_loss.mean().detach()
         return loss, None, None
 
     def log(self, logs: dict[str, float], start_time: Optional[float] = None) -> None:
