--- src/unexploitable_search/train/trainers/game_trainer/game_trainer_hard.py
@@ -1,3878 +1,0 @@
-# type: ignore
-#####################################################################
-# THIS FILE IS A COPY OF THE TRL GRPO TRAINER FILE                  #
-# WITH THE GAME LOGIC ADDED.                                     #
-# DUE TO UNSLOTH'S MONKEY-PATCHING STRATEGY, INHERITANCE IS NOT     #
-# A STABLE SOLUTION.                                                #
-#####################################################################
-
-# Copyright 2020-2025 The HuggingFace Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import inspect
-import os
-import re
-import textwrap
-from collections import defaultdict, deque
-from contextlib import nullcontext
-from functools import partial
-from pathlib import Path
-from typing import Any, Callable, Optional, Union, List, Dict, cast
-
-import datasets
-import torch
-import torch.utils.data
-import transformers
-from accelerate import logging
-from accelerate.utils import (
-    broadcast_object_list,
-    gather,
-    gather_object,
-    is_peft_model,
-    set_seed,
-)
-from datasets import Dataset, IterableDataset
-from torch import nn
-from peft import PeftMixedModel, PeftModel
-from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
-from torch.utils.data import DataLoader, Sampler
-from transformers import (
-    AutoConfig,
-    AutoModelForSequenceClassification,
-    AutoProcessor,
-    AutoTokenizer,
-    GenerationConfig,
-    PreTrainedModel,
-    PreTrainedTokenizerBase,
-    ProcessorMixin,
-    TrainerCallback,
-    is_wandb_available,
-)
-from transformers.trainer_utils import seed_worker
-from transformers.training_args import DistributedType, OptimizerNames
-from transformers.utils import (
-    is_datasets_available,
-    is_sagemaker_mp_enabled,
-    is_torch_hpu_available,
-    is_torch_mlu_available,
-    is_torch_mps_available,
-    is_torch_musa_available,
-    is_torch_npu_available,
-    is_torch_xpu_available,
-    is_peft_available,
-    is_flash_attn_2_available,
-    is_rich_available,
-)
-
-from unexploitable_search.train.trainers.game_trainer.prompts import (
-    BOB_SYSTEM_PROMPT,
-    CHARLIE_SYSTEM_PROMPT,
-    PROPERTY_VERIFICATION_SYSTEM_PROMPT,
-)
-
-from trl.data_utils import (
-    apply_chat_template,
-    is_conversational,
-    maybe_apply_chat_template,
-    prepare_multimodal_messages,
-)
-from trl.extras.profiling import profiling_context, profiling_decorator
-from trl.extras.vllm_client import VLLMClient
-from trl.import_utils import is_liger_kernel_available, is_vllm_available
-from trl.models import (
-    prepare_deepspeed,
-    prepare_fsdp,
-    prepare_peft_model,
-    unwrap_model_for_generation,
-)
-from trl.models.utils import _ForwardRedirection
-from trl.trainer.base_trainer import BaseTrainer
-from trl.trainer.callbacks import SyncRefModelCallback
-from trl.trainer.grpo_config import GRPOConfig
-from trl.trainer.utils import (
-    RepeatSampler,
-    disable_dropout_in_model,
-    ensure_master_addr_port,
-    entropy_from_logits,
-    identity,
-    nanmax,
-    nanmin,
-    nanstd,
-    pad,
-    print_prompt_completions_sample,
-    selective_log_softmax,
-    shuffle_sequence_dict,
-    split_pixel_values_by_grid,
-    split_tensor_dict,
-    truncate_with_protected_tokens,
-    unsplit_pixel_values_by_grid,
-)
-
-
-if is_peft_available():
-    from peft import PeftConfig, PeftModel
-
-if is_liger_kernel_available():
-    from liger_kernel.chunked_loss import LigerFusedLinearGRPOLoss
-
-if is_vllm_available():
-    from vllm import LLM, SamplingParams
-    from vllm.sampling_params import GuidedDecodingParams
-
-if is_wandb_available():
-    import wandb
-
-if is_sagemaker_mp_enabled():
-    from smdistributed.modelparallel import __version__ as SMP_VERSION
-
-    IS_SAGEMAKER_MP_POST_1_10 = version.parse(SMP_VERSION) >= version.parse("1.10")  # noqa: F821
-
-    from trl.trainer.trainer_pt_utils import smp_forward_backward
-else:
-    IS_SAGEMAKER_MP_POST_1_10 = False
-
-
-logger = logging.get_logger(__name__)
-
-# What we call a reward function is a callable that takes a list of prompts and completions and returns a list of
-# rewards. When it's a string, it's a model ID, so it's loaded as a pretrained model.
-RewardFunc = Union[str, PreTrainedModel, Callable[[list, list], list[float]]]
-
-
-class GRPOHardGameTrainer(BaseTrainer):
-    """
-    A GRPOTrainer specialized for the two-player game mitigation strategy.
-
-    This trainer enforces that the model must be a PEFT model and provides
-    configurable prompts for Bob and the property checker.
-    During initialization, it creates two additional LoRA adapters using
-    the same config as existing adapters:
-    - 'property_checker': For property verification (not trained)
-    - 'bob': For Bob's completions (will be trained)
-
-    Example:
-
-    ```python
-    from datasets import load_dataset
-    from trl import GRPOTrainer
-
-    dataset = load_dataset("trl-lib/tldr", split="train")
-
-
-    def reward_func(completions, **kwargs):
-        # Dummy reward function that rewards completions with more unique letters.
-        return [float(len(set(completion))) for completion in completions]
-
-
-    trainer = GRPOTrainer(
-        model="Qwen/Qwen2-0.5B-Instruct",
-        reward_funcs=reward_func,
-        train_dataset=dataset,
-    )
-
-    trainer.train()
-    ```
-
-    Args:
-        model (`Union[str, PreTrainedModel]`):
-            Model to be trained. Can be either:
-
-            - A string, being the *model id* of a pretrained model hosted inside a model repo on huggingface.co, or a
-              path to a *directory* containing model weights saved using
-              [`~transformers.PreTrainedModel.save_pretrained`], e.g., `'./my_model_directory/'`. The model is loaded
-              using [`~transformers.AutoModelForCausalLM.from_pretrained`] with the keyword arguments in
-              `args.model_init_kwargs`.
-            - A [`~transformers.PreTrainedModel`] object. Only causal language models are supported.
-        reward_funcs (`Union[RewardFunc, list[RewardFunc]]`):
-            Reward functions to be used for computing the rewards. To compute the rewards, we call all the reward
-            functions with the prompts and completions and sum the rewards. Can be either:
-
-            - A single reward function, such as:
-                - A string: The *model ID* of a pretrained model hosted inside a model repo on huggingface.co, or a
-                path to a *directory* containing model weights saved using
-                [`~transformers.PreTrainedModel.save_pretrained`], e.g., `'./my_model_directory/'`. The model is loaded
-                using [`~transformers.AutoModelForSequenceClassification.from_pretrained`] with `num_labels=1` and the
-                keyword arguments in `args.model_init_kwargs`.
-                - A [`~transformers.PreTrainedModel`] object: Only sequence classification models are supported.
-                - A custom reward function: The function is provided with the prompts and the generated completions,
-                  plus any additional columns in the dataset. It should return a list of rewards. Custom reward
-                  functions can also return `None` when the reward is not applicable to those samples. This is useful
-                  for multi-task training where different reward functions apply to different types of samples. When a
-                  reward function returns `None` for a sample, that reward function is excluded from the reward
-                  calculation for that sample. For more details, see [Using a custom reward
-                  function](#using-a-custom-reward-function).
-
-                  The trainer's state is also passed to the reward function. The trainer's state is an instance of
-                  [`~transformers.TrainerState`] and can be accessed by accessing the `trainer_state` argument to the
-                  reward function's signature.
-            - A list of reward functions, where each item can independently be any of the above types. Mixing different
-            types within the list (e.g., a string model ID and a custom reward function) is allowed.
-        args ([`GRPOConfig`], *optional*):
-            Configuration for this trainer. If `None`, a default configuration is used.
-        train_dataset ([`~datasets.Dataset`] or [`~datasets.IterableDataset`]):
-            Dataset to use for training. It must include a column `"prompt"`. Any additional columns in the dataset is
-            ignored. The format of the samples can be either:
-
-            - [Standard](dataset_formats#standard): Each sample contains plain text.
-            - [Conversational](dataset_formats#conversational): Each sample contains structured messages (e.g., role
-              and content).
-        eval_dataset ([`~datasets.Dataset`], [`~datasets.IterableDataset`] or `dict[str, Union[Dataset, IterableDataset]]`):
-            Dataset to use for evaluation. It must meet the same requirements as `train_dataset`.
-        processing_class ([`~transformers.PreTrainedTokenizerBase`], [`~transformers.ProcessorMixin`], *optional*):
-            Processing class used to process the data. The padding side must be set to "left". If `None`, the
-            processing class is loaded from the model's name with [`~transformers.AutoProcessor.from_pretrained`]. A
-            padding token, `tokenizer.pad_token`, must be set. If the processing class has not set a padding token,
-            `tokenizer.eos_token` will be used as the default.
-        reward_processing_classes (`Union[PreTrainedTokenizerBase, list[PreTrainedTokenizerBase]]`, *optional*):
-            Processing classes corresponding to the reward functions specified in `reward_funcs`. Can be either:
-
-            - A single processing class: Used when `reward_funcs` contains only one reward function.
-            - A list of processing classes: Must match the order and length of the reward functions in `reward_funcs`.
-            If set to `None`, or if an element of the list corresponding to a [`~transformers.PreTrainedModel`] is
-            `None`, the tokenizer for the model is automatically loaded using
-            [`~transformers.AutoTokenizer.from_pretrained`]. For elements in `reward_funcs` that are custom reward
-            functions (not [`~transformers.PreTrainedModel`]), the corresponding entries in `reward_processing_classes`
-            are ignored.
-        callbacks (list of [`~transformers.TrainerCallback`], *optional*):
-            List of callbacks to customize the training loop. Will add those to the list of default callbacks detailed
-            in [here](https://huggingface.co/docs/transformers/main_classes/callback).
-
-            If you want to remove one of the default callbacks used, use the [`~transformers.Trainer.remove_callback`]
-            method.
-        optimizers (`tuple[torch.optim.Optimizer, torch.optim.lr_scheduler.LambdaLR]`, *optional*, defaults to `(None, None)`):
-            A tuple containing the optimizer and the scheduler to use. Will default to an instance of [`AdamW`] on your
-            model and a scheduler given by [`get_linear_schedule_with_warmup`] controlled by `args`.
-        peft_config ([`~peft.PeftConfig`], *optional*):
-            PEFT configuration used to wrap the model. If `None`, the model is not wrapped.
-    """
-
-    DEFAULT_BOB_PROMPT = BOB_SYSTEM_PROMPT
-    DEFAULT_CHARLIE_PROMPT = CHARLIE_SYSTEM_PROMPT
-    DEFAULT_PROPERTY_CHECKER_PROMPT = PROPERTY_VERIFICATION_SYSTEM_PROMPT
-    _tag_names = ["trl", "grpo"]
-    _name = "GRPO"
-    _paper = {
-        "title": "DeepSeekMath: Pushing the Limits of Mathematical Reasoning in Open Language Models",
-        "id": "2402.03300",
-        # docstyle-ignore
-        "citation": textwrap.dedent("""\
-            @article{shao2024deepseekmath,
-                title        = {{DeepSeekMath: Pushing the Limits of Mathematical Reasoning in Open Language Models}},
-                author       = {Zhihong Shao and Peiyi Wang and Qihao Zhu and Runxin Xu and Junxiao Song and Mingchuan Zhang and Y. K. Li and Y. Wu and Daya Guo},
-                year         = 2024,
-                eprint       = {arXiv:2402.03300},
-            }
-            """),
-    }
-
-    def __init__(
-        self,
-        model: Union[str, PreTrainedModel],
-        reward_funcs: Union[RewardFunc, list[RewardFunc]],
-        args: Optional[GRPOConfig] = None,
-        train_dataset: Optional[Union[Dataset, IterableDataset]] = None,
-        eval_dataset: Optional[
-            Union[Dataset, IterableDataset, dict[str, Union[Dataset, IterableDataset]]]
-        ] = None,
-        processing_class: Optional[
-            Union[PreTrainedTokenizerBase, ProcessorMixin]
-        ] = None,
-        reward_processing_classes: Optional[
-            Union[PreTrainedTokenizerBase, list[PreTrainedTokenizerBase]]
-        ] = None,
-        callbacks: Optional[list[TrainerCallback]] = None,
-        optimizers: tuple[
-            Optional[torch.optim.Optimizer], Optional[torch.optim.lr_scheduler.LambdaLR]
-        ] = (None, None),
-        peft_config: Optional["PeftConfig"] = None,
-        bob_prompt: Optional[str] = None,
-        charlie_prompt: Optional[str] = None,
-        property_checker_prompt: Optional[str] = None,
-        game_coefficient: float = 0.2,
-        tolerance: float = 0.1,
-        solution_preprocessor: Optional[Callable[[str], str]] = None,
-        preprocess_reference_solutions: bool = False,
-        max_property_length: int = 20,
-        num_references: int = 1,
-        game_score_weights: list[float] | None = None,
-        train_adversary: bool = True,
-    ):
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        has_preprocessor = solution_preprocessor is not None
-        preprocessor_is_callable = callable(solution_preprocessor)
-        if has_preprocessor and not preprocessor_is_callable:
-            raise ValueError("solution_preprocessor must be a callable function")
-        if preprocess_reference_solutions and not has_preprocessor:
-            raise ValueError(
-                "solution_preprocessor must be provided "
-                "if preprocess_reference_solutions is True"
-            )
-
-        # Validate num_references
-        if num_references < 1:
-            raise ValueError(f"num_references must be at least 1, got {num_references}")
-
-        # Validate that model is a PEFT model
-        if not isinstance(model, (PeftModel, PeftMixedModel)):
-            raise TypeError(
-                "GRPOGameTrainer requires a PEFT model"
-                " (PeftModel or PeftMixedModel). "
-                f"Got {type(model).__name__} instead. "
-                "Please wrap your model with PEFT before "
-                "passing it to this trainer."
-            )
-
-        # Set prompts with defaults
-        self.bob_prompt = (
-            bob_prompt if bob_prompt is not None else self.DEFAULT_BOB_PROMPT
-        )
-        self.charlie_prompt = (
-            charlie_prompt
-            if charlie_prompt is not None
-            else self.DEFAULT_CHARLIE_PROMPT
-        )
-        self.property_checker_prompt = (
-            property_checker_prompt
-            if property_checker_prompt is not None
-            else self.DEFAULT_PROPERTY_CHECKER_PROMPT
-        )
-
-        # Store game parameters
-        self.game_coefficient = game_coefficient
-        self.tolerance = tolerance
-        self.solution_preprocessor = solution_preprocessor
-        self.preprocess_reference_solutions = preprocess_reference_solutions
-        self.max_property_length = max_property_length
-        self.num_references = num_references
-        self.game_score_weights = game_score_weights
-        self.train_adversary = train_adversary
-
-        # Store PEFT model reference
-        self._peft_model = model
-
-        # Extract the existing LoRA config from the first adapter
-        # This assumes the model already has at least one adapter
-        existing_adapter_names = list(model.peft_config.keys())
-        if not existing_adapter_names:
-            raise ValueError(
-                "The provided PEFT model has no adapters. "
-                "Please ensure the model is initialized "
-                "with at least one adapter."
-            )
-
-        # Get the config from the first adapter
-        base_lora_config = model.peft_config[existing_adapter_names[0]]
-
-        # Create new adapters with the same configuration
-        # Property checker adapter - will be used for property verification
-        self.property_checker_adapter_name = "property_checker"
-        model.add_adapter(self.property_checker_adapter_name, base_lora_config)
-
-        # Bob's adapter - will be used for Bob's generations (will be trained)
-        self.bob_adapter_name = "bob"
-        model.add_adapter(self.bob_adapter_name, base_lora_config)
-
-        # Charlie's adapter - will be used for Charlie's generations (will be trained)
-        self.charlie_adapter_name = "charlie"
-        model.add_adapter(self.charlie_adapter_name, base_lora_config)
-
-        # Only enable gradients for adversary adapters if training them
-        if self.train_adversary:
-            for name, param in model.named_parameters():
-                if self.bob_adapter_name in name or self.charlie_adapter_name in name:
-                    param.requires_grad = True
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        # Args
-        if args is None:
-            model_name = model if isinstance(model, str) else model.config._name_or_path
-            model_name = model_name.split("/")[-1]
-            args = GRPOConfig(f"{model_name}-GRPO")
-
-        # Models
-        # Trained model
-        model_init_kwargs = args.model_init_kwargs or {}
-        if isinstance(model, str):
-            model_id = model
-            dtype = model_init_kwargs.get("dtype")
-            if isinstance(dtype, torch.dtype) or dtype == "auto" or dtype is None:
-                pass  # dtype is already a torch.dtype or "auto" or None
-            elif isinstance(dtype, str):  # it's a str, but not "auto"
-                dtype = getattr(torch, dtype)
-                model_init_kwargs["dtype"] = dtype
-            else:
-                raise ValueError(
-                    "Invalid `dtype` passed to `GRPOConfig`. Expected either 'auto' or a string representing "
-                    f"a `torch.dtype` (e.g., 'float32'), but got {dtype}."
-                )
-            # Disable caching if gradient checkpointing is enabled (not supported)
-            config = AutoConfig.from_pretrained(model_id)
-            architecture = getattr(transformers, config.architectures[0])
-            model = architecture.from_pretrained(model_id, **model_init_kwargs)
-        else:
-            model_id = model.config._name_or_path
-            if args.model_init_kwargs is not None:
-                logger.warning(
-                    "You passed `model_init_kwargs` to the `GRPOConfig`, but your model is already instantiated. "
-                    "The `model_init_kwargs` will be ignored."
-                )
-
-        # Some models (SmolVLM/Idefics3) don't support `logits_to_keep` argument and error out if we pass it
-        # Inspect the forward method before we wrap the model with PEFT
-        self.model_kwarg_keys = (
-            inspect.signature(model.forward).parameters.keys()
-            if not hasattr(model, "get_base_model")
-            else inspect.signature(model.get_base_model().forward).parameters.keys()
-        )
-
-        if peft_config is not None or (
-            is_peft_available() and isinstance(model, PeftModel)
-        ):
-            model = prepare_peft_model(model, peft_config, args)
-
-        # Processing class
-        if processing_class is None:
-            processing_class = AutoProcessor.from_pretrained(model.config._name_or_path)
-
-        # Handle pad token for processors or tokenizers
-        if isinstance(processing_class, ProcessorMixin):
-            tokenizer = processing_class.tokenizer
-        elif isinstance(processing_class, PreTrainedTokenizerBase):
-            tokenizer = processing_class
-        else:
-            raise TypeError(
-                "The `processing_class` must be either a `PreTrainedTokenizerBase` or a `ProcessorMixin`"
-            )
-
-        if tokenizer.pad_token is None:
-            tokenizer.pad_token = tokenizer.eos_token
-
-        self.pad_token = tokenizer.pad_token
-        self.pad_token_id = tokenizer.pad_token_id
-        self.eos_token_id = tokenizer.eos_token_id
-        self.image_token = getattr(processing_class, "image_token", None)
-        self.image_token_id = getattr(processing_class, "image_token_id", None)
-        self.vision_start_token_id = getattr(
-            model.config, "vision_start_token_id", None
-        )
-        self.vision_end_token_id = getattr(model.config, "vision_end_token_id", None)
-
-        # Reward functions
-        if not isinstance(reward_funcs, list):
-            reward_funcs = [reward_funcs]
-        self.reward_func_names = []
-        for i, reward_func in enumerate(reward_funcs):
-            if isinstance(reward_func, str):
-                reward_funcs[i] = AutoModelForSequenceClassification.from_pretrained(
-                    reward_func, num_labels=1, **model_init_kwargs
-                )
-            if isinstance(
-                reward_funcs[i], nn.Module
-            ):  # Use Module over PretrainedModel for compat w/ compiled models
-                self.reward_func_names.append(
-                    reward_funcs[i].config._name_or_path.split("/")[-1]
-                )
-            else:
-                self.reward_func_names.append(reward_funcs[i].__name__)
-        self.reward_funcs = reward_funcs
-
-        if self.game_score_weights is not None:
-            if len(self.game_score_weights) != len(reward_funcs):
-                raise ValueError(
-                    f"Number of game score weights ({len(self.game_score_weights)}) must match number of reward "
-                    f"functions ({len(reward_funcs)})."
-                )
-            self.game_score_weights = torch.tensor(
-                self.game_score_weights, dtype=torch.float32
-            )
-        else:
-            self.game_score_weights = torch.ones(len(reward_funcs), dtype=torch.float32)
-
-        # Reward weights
-        if args.reward_weights is not None:
-            if len(args.reward_weights) != len(reward_funcs):
-                raise ValueError(
-                    f"Number of reward weights ({len(args.reward_weights)}) must match number of reward "
-                    f"functions ({len(reward_funcs)})"
-                )
-            self.reward_weights = torch.tensor(args.reward_weights, dtype=torch.float32)
-        else:
-            self.reward_weights = torch.ones(len(reward_funcs), dtype=torch.float32)
-
-        # Reward processing class
-        if reward_processing_classes is None:
-            reward_processing_classes = [None] * len(reward_funcs)
-        elif not isinstance(reward_processing_classes, list):
-            reward_processing_classes = [reward_processing_classes]
-        if len(reward_processing_classes) != len(reward_funcs):
-            raise ValueError(
-                f"The number of reward processing classes ({len(reward_processing_classes)}) must match the number of "
-                f"reward functions ({len(reward_funcs)})."
-            )
-
-        for i, (reward_processing_class, reward_func) in enumerate(
-            zip(reward_processing_classes, reward_funcs)
-        ):
-            if isinstance(reward_func, PreTrainedModel):
-                if reward_processing_class is None:
-                    reward_processing_class = AutoTokenizer.from_pretrained(
-                        reward_func.config._name_or_path
-                    )
-                if reward_processing_class.pad_token_id is None:
-                    reward_processing_class.pad_token = (
-                        reward_processing_class.eos_token
-                    )
-                # The reward model computes the reward for the latest non-padded token in the input sequence.
-                # So it's important to set the pad token ID to the padding token ID of the processing class.
-                reward_func.config.pad_token_id = reward_processing_class.pad_token_id
-                reward_processing_classes[i] = reward_processing_class
-
-        self.reward_processing_classes = reward_processing_classes
-
-        # Training arguments
-        self.max_prompt_length = args.max_prompt_length
-        self.max_completion_length = (
-            args.max_completion_length
-        )  # = |o_i| in the GRPO paper
-        self.num_generations = args.num_generations  # = G in the GRPO paper
-        self.num_generations_eval = args.num_generations_eval
-        self.temperature = args.temperature
-        self.top_p = args.top_p
-        self.top_k = args.top_k
-        self.min_p = args.min_p
-        self.repetition_penalty = args.repetition_penalty
-        self.use_transformers_paged = args.use_transformers_paged
-        self.use_vllm = args.use_vllm
-        self.vllm_mode = args.vllm_mode
-        self.vllm_gpu_memory_utilization = (
-            args.vllm_gpu_memory_utilization
-        )  # only applies to colocation mode
-        self.vllm_tensor_parallel_size = (
-            args.vllm_tensor_parallel_size
-        )  # only applies to colocation mode
-        self.vllm_importance_sampling_correction = (
-            args.vllm_importance_sampling_correction
-        )
-        self.vllm_importance_sampling_cap = args.vllm_importance_sampling_cap
-        self.use_liger_loss = args.use_liger_loss
-        self.loss_type = args.loss_type
-        self.scale_rewards = args.scale_rewards
-        self.importance_sampling_level = args.importance_sampling_level
-        self.mask_truncated_completions = args.mask_truncated_completions
-        self.top_entropy_quantile = args.top_entropy_quantile
-        if self.use_liger_loss and self.top_entropy_quantile < 1.0:
-            raise NotImplementedError(
-                "Liger Kernels don't currently support masking token positions based on entropy."
-            )
-        if self.use_liger_loss and not self.importance_sampling_level == "token":
-            raise NotImplementedError(
-                "Liger Kernels currently only support token-level importance sampling. Please set"
-                "`importance_sampling_level` to 'token'."
-            )
-
-        # Datasets
-        self.shuffle_dataset = args.shuffle_dataset
-
-        if (
-            isinstance(train_dataset, IterableDataset)
-            or isinstance(eval_dataset, IterableDataset)
-            or (
-                isinstance(eval_dataset, dict)
-                and any(isinstance(ds, IterableDataset) for ds in eval_dataset.values())
-            )
-        ):
-            # See https://github.com/huggingface/trl/issues/3213
-            raise NotImplementedError(
-                "Iterable datasets are not yet supported in GRPOTrainer. Please use a standard dataset instead."
-            )
-
-        # Multi-step
-        self.num_iterations = args.num_iterations  # = 𝜇 in the GRPO paper
-        self.epsilon_low = args.epsilon
-        self.epsilon_high = (
-            args.epsilon_high if args.epsilon_high is not None else args.epsilon
-        )
-        # Tracks the number of iterations (forward + backward passes), including those within a grad accum cycle
-        self._step = 0
-        # Buffer the batch to reuse generated outputs across multiple updates. For more details, see
-        # `_get_train_sampler` and `_prepare_inputs`.
-        self._buffered_inputs = None
-
-        # The trainer estimates the number of FLOPs (floating-point operations) using the number of elements in the
-        # input tensor associated with the key "input_ids". However, in GRPO, the sampled data does not include the
-        # "input_ids" key. Instead, the available keys is "prompt". As a result, the trainer issues the warning:
-        # "Could not estimate the number of tokens of the input, floating-point operations will not be computed." To
-        # suppress this warning, we set the "estimate_tokens" key in the model's "warnings_issued" dictionary to True.
-        # This acts as a flag to indicate that the warning has already been issued.
-        model.warnings_issued["estimate_tokens"] = True
-
-        super().__init__(
-            model=model,
-            args=args,
-            data_collator=identity,  # No data collation is needed in GRPO
-            train_dataset=train_dataset,
-            eval_dataset=eval_dataset,
-            processing_class=processing_class,
-            callbacks=callbacks,
-            optimizers=optimizers,
-            # In Trainer, `training_step` scales the loss by `gradient_accumulation_steps` only if `compute_loss_func`
-            # is None. For DAPO, loss scaling instead depends on the total number of completions tokens across the
-            # global accumulated batch. To control scaling ourselves, we must disable Trainer’s built-in scaling. The
-            # simplest (though a bit hacky) way is to set `compute_loss_func` to any non-None value, which bypasses
-            # that behavior without rewriting `training_step`.
-            compute_loss_func="non-None value to disable scaling",
-        )
-
-        # Reference model
-        self.beta = args.beta
-        if self.beta == 0.0:
-            # If beta is 0.0, the reference model is not needed
-            self.ref_model = None
-        elif is_peft_model(model):
-            # If PEFT is used, the reference model is not needed since the adapter can be disabled
-            # to revert to the initial model.
-            self.ref_model = None
-        else:
-            # For deepspeed, fsdp or non-distributed models, create a reference model from scratch
-            config = AutoConfig.from_pretrained(model_id)
-            architecture = getattr(transformers, config.architectures[0])
-            self.ref_model = architecture.from_pretrained(model_id, **model_init_kwargs)
-
-        # Disable dropout in the models
-        if args.disable_dropout:
-            disable_dropout_in_model(model)
-            if self.ref_model is not None:
-                disable_dropout_in_model(self.ref_model)
-
-        # Liger loss
-        if self.use_liger_loss:
-            if not is_liger_kernel_available():
-                raise ImportError(
-                    "Liger is required to use `liger_loss` as the GRPO loss. Run `pip install liger-kernel`."
-                )
-            # redirect the model.module forward to the model forward to ensure pre-forward hooks are called
-            self._forward_redirection = _ForwardRedirection()
-
-            self.liger_grpo_loss = LigerFusedLinearGRPOLoss(
-                beta=self.beta,
-                epsilon_low=self.epsilon_low,
-                epsilon_high=self.epsilon_high,
-                temperature=self.temperature,
-                use_ref_model=self.beta != 0.0,
-                loss_type=self.loss_type,
-                max_completion_length=self.max_completion_length,
-            )
-
-        # Initialize the metrics
-        self._metrics = {"train": defaultdict(list), "eval": defaultdict(list)}
-        self._total_train_tokens = 0
-        self.log_completions = args.log_completions
-        self.wandb_log_unique_prompts = args.wandb_log_unique_prompts
-        self.num_completions_to_print = args.num_completions_to_print
-        # Keep logs sized to the generation batch to record only outputs from the latest model update.
-        self._logs = {
-            "images": deque(maxlen=args.generation_batch_size),
-            "prompt": deque(maxlen=args.generation_batch_size),
-            "completion": deque(maxlen=args.generation_batch_size),
-            "rewards": defaultdict(lambda: deque(maxlen=args.generation_batch_size)),
-            "advantages": deque(maxlen=args.generation_batch_size),
-        }
-
-        # Ensure each process receives a unique seed to prevent duplicate completions when generating with
-        # transformers if num_generations exceeds per_device_train_batch_size. We could skip it if we use vLLM, but
-        # it's safer to set it in all cases.
-        set_seed(args.seed, device_specific=True)
-
-        if self.use_vllm:
-            if not is_vllm_available():
-                raise ImportError(
-                    "vLLM is not available and `use_vllm` is set to True. Please install vLLM with "
-                    "`pip install trl[vllm]` to use it."
-                )
-
-            if self.vllm_mode == "server":
-                if self.accelerator.is_main_process:
-                    if args.vllm_server_base_url is not None:
-                        base_url = args.vllm_server_base_url
-                    else:
-                        base_url = (
-                            f"http://{args.vllm_server_host}:{args.vllm_server_port}"
-                        )
-                    self.vllm_client = VLLMClient(
-                        base_url=base_url, connection_timeout=args.vllm_server_timeout
-                    )
-                    self.vllm_client.init_communicator(
-                        device=torch.cuda.current_device()
-                    )
-
-            elif self.vllm_mode == "colocate":
-                # Make sure vllm_tensor_parallel_size group size evenly divides the world size - each group should have
-                # the same number of ranks
-                if (
-                    not self.accelerator.num_processes % self.vllm_tensor_parallel_size
-                    == 0
-                ):
-                    raise ValueError(
-                        f"vllm_tensor_parallel_size ({self.vllm_tensor_parallel_size}) must divide world size "
-                        f"({self.accelerator.num_processes}) evenly."
-                    )
-
-                if self.vllm_tensor_parallel_size > 1:
-                    # Create subgroups of ranks for TP, each group with `vllm_tensor_parallel_size` ranks.
-                    # For example, if world_size=8 and vllm_tensor_parallel_size=2 → groups: [0,1], [2,3], [4,5], [6,7]
-                    self.tp_group, _ = torch.distributed.new_subgroups_by_enumeration(
-                        [
-                            list(
-                                range(
-                                    i * self.vllm_tensor_parallel_size,
-                                    (i + 1) * self.vllm_tensor_parallel_size,
-                                )
-                            )
-                            for i in range(
-                                self.accelerator.num_processes
-                                // self.vllm_tensor_parallel_size
-                            )
-                        ]
-                    )
-
-                # vLLM requires the environment variables to be set for distributed training.
-                os.environ["RANK"] = str(self.accelerator.process_index)
-                os.environ["LOCAL_RANK"] = str(self.accelerator.local_process_index)
-                os.environ["WORLD_SIZE"] = str(self.accelerator.num_processes)
-                # Ensure distributed rendezvous variables are set without colliding across concurrent runs
-                ensure_master_addr_port()
-
-                if (
-                    self.max_prompt_length is not None
-                    and self.max_completion_length is not None
-                ):
-                    max_model_len = self.max_prompt_length + self.max_completion_length
-                else:
-                    max_model_len = None
-                self.llm = LLM(
-                    model=model.name_or_path,
-                    tensor_parallel_size=args.vllm_tensor_parallel_size,
-                    gpu_memory_utilization=self.vllm_gpu_memory_utilization,
-                    max_num_seqs=self.args.per_device_train_batch_size
-                    * self.vllm_tensor_parallel_size
-                    * self.args.steps_per_generation,
-                    max_model_len=max_model_len,
-                    distributed_executor_backend="external_launcher",
-                    # Feed identical seed for tp groups to ensure sampling results are the same across workers
-                    seed=self.accelerator.process_index
-                    // self.vllm_tensor_parallel_size,
-                    # Latest vLLM v1 memory profiler is misled by the high default value (i.e., 32768) - thinking there's not enough memory
-                    max_num_batched_tokens=4096,
-                    model_impl=self.args.vllm_model_impl,
-                    enable_sleep_mode=self.args.vllm_enable_sleep_mode,
-                    # Important so temperature scaling/logit tweaking affects the TIS log probs
-                    logprobs_mode="processed_logprobs",
-                )
-                if self.args.vllm_enable_sleep_mode:
-                    self.llm.sleep(level=1)
-            else:
-                raise ValueError(
-                    f"vllm_mode must be either 'server' or 'colocate', got '{self.vllm_mode}'."
-                )
-
-            # vLLM specific sampling arguments
-            self.guided_decoding_regex = args.vllm_guided_decoding_regex
-
-            self._last_loaded_step = (
-                -1
-            )  # tag to avoid useless loading during grad accumulation
-
-            # When using vLLM, the main process is responsible for loading the model weights. This can cause process
-            # desynchronization and seems to lead to DeepSpeed hanging during initialization. To prevent this, we
-            # synchronize all processes after vLLM has been fully initialized.
-            self.accelerator.wait_for_everyone()
-        else:
-            generation_kwargs = {
-                "max_new_tokens": self.max_completion_length,
-                "do_sample": True,
-                "pad_token_id": tokenizer.pad_token_id,
-                "bos_token_id": tokenizer.bos_token_id,
-                "eos_token_id": tokenizer.eos_token_id,
-                "temperature": self.temperature,
-                "top_p": self.top_p,
-                "top_k": self.top_k,
-                "min_p": self.min_p,
-                "repetition_penalty": self.repetition_penalty,
-                "cache_implementation": args.cache_implementation,
-            }
-            if args.generation_kwargs is not None:
-                generation_kwargs.update(args.generation_kwargs)
-            self.generation_config = GenerationConfig(**generation_kwargs)
-
-        # Gradient accumulation requires scaled loss. Normally, loss scaling in the parent class depends on whether the
-        # model accepts loss-related kwargs. Since we compute our own loss, this check is irrelevant. We set
-        # self.model_accepts_loss_kwargs to False to enable scaling.
-        self.model_accepts_loss_kwargs = False
-
-        # Add tags to the model
-        self.model.add_model_tags(self._tag_names)
-
-        if self.ref_model is not None:
-            if self.is_deepspeed_enabled:
-                self.ref_model = prepare_deepspeed(self.ref_model, self.accelerator)
-            elif self.is_fsdp_enabled:
-                self.ref_model = prepare_fsdp(self.ref_model, self.accelerator)
-            else:
-                self.ref_model = self.accelerator.prepare_model(
-                    self.ref_model, evaluation_mode=True
-                )
-
-        if args.sync_ref_model:
-            self.add_callback(
-                SyncRefModelCallback(
-                    ref_model=self.ref_model, accelerator=self.accelerator
-                )
-            )
-
-        for i, reward_func in enumerate(self.reward_funcs):
-            if isinstance(reward_func, PreTrainedModel):
-                if self.is_deepspeed_enabled:
-                    self.reward_funcs[i] = prepare_deepspeed(
-                        reward_func, self.accelerator
-                    )
-                else:
-                    # set device placement to True to make `prepare_model` move `reward_func` to device when using fsdp
-                    self.reward_funcs[i] = self.accelerator.prepare_model(
-                        reward_func, evaluation_mode=True, device_placement=True
-                    )
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        # Storage for reference completions
-        # (n+1-th completion per unique prompt)
-        self._reference_prompt_ids: List[List[int]] = []
-        self._reference_completion_ids: List[List[int]] = []
-        self._reference_logprobs: Optional[List[Any]] = None
-        self._reference_prompts: List[
-            Any
-        ] = []  # Original prompts for reference completions
-
-        # Storage for Bob's completions
-        self._bob_prompt_ids: List[List[int]] = []
-        self._bob_completion_ids: List[List[int]] = []
-        self._bob_logprobs: Optional[List[Any]] = None
-
-        # Storage for Charlie's completions
-        self._charlie_prompt_ids: List[List[int]] = []
-        self._charlie_completion_ids: List[List[int]] = []
-        self._charlie_logprobs: Optional[List[Any]] = None
-
-        # Add game as an additional reward function
-        self.reward_func_names.append("game_reward")
-        self.reward_weights = torch.cat(
-            [
-                self.reward_weights,
-                torch.tensor([self.game_coefficient], dtype=torch.float32),
-            ]
-        )
-
-        # Initialize Bob-specific log entries
-        self._logs["bob_prompt"] = deque(maxlen=args.generation_batch_size)
-        self._logs["bob_completion"] = deque(maxlen=args.generation_batch_size)
-
-        # Initialize Charlie-specific log entries
-        self._logs["charlie_prompt"] = deque(maxlen=args.generation_batch_size)
-        self._logs["charlie_completion"] = deque(maxlen=args.generation_batch_size)
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-    @property
-    def reference_logprobs(self) -> Optional[List[Any]]:
-        """Access to the stored reference logprobs."""
-        return self._reference_logprobs
-
-    def get_bob_prompt(self, property_description: str) -> str:
-        """
-        Format Bob's prompt with a specific property description.
-
-        Args:
-            property_description: The property that Bob should avoid
-
-        Returns:
-            Formatted prompt string
-        """
-        return self.bob_prompt.format(property_description=property_description)
-
-    def get_property_checker_prompt(self) -> str:
-        """
-        Get the property checker system prompt.
-
-        Returns:
-            Property checker prompt string
-        """
-        return self.property_checker_prompt
-
-    def _format_player_prompts(
-        self,
-        original_prompts_batch: List[List[str]],
-        reference_solutions_batch: List[List[str]],
-        system_prompt: str,
-    ) -> List[List[Dict[str, Any]]]:
-        """
-        Format prompts for a player (Bob or Charlie) given the original prompts,
-        reference solutions, and system prompt.
-
-        Args:
-            original_prompts_batch: List of original prompt messages (chat format)
-            reference_solutions_batch: List of reference solution code strings
-            system_prompt: System prompt for the player
-
-        Returns:
-            List of formatted prompts (chat format)
-        """
-        player_prompts = []
-
-        for original_prompts, reference_solutions in zip(
-            original_prompts_batch, reference_solutions_batch
-        ):
-            # Extract the user content from the original prompt
-            if isinstance(original_prompts, list):
-                # The prompt is the same for each batched reference solution
-                original_prompt = original_prompts[0]
-                # Find the user message content
-                user_content = None
-                for message in original_prompt:
-                    if isinstance(message, dict) and message.get("role") == "user":
-                        user_content = message.get("content", "")
-                        break
-
-                if user_content is None:
-                    user_content = str(original_prompt)
-            else:
-                user_content = original_prompts
-
-            # Format player's user message with problem and reference solution
-            player_user_message = f"PROBLEM:\n\n{user_content}\n\n"
-            if self.num_references > 0:
-                player_user_message += "REFERENCE SOLUTION(S):\n\n"
-                for reference_solution in reference_solutions:
-                    player_user_message += (
-                        f"<solution>\n{reference_solution}\n</solution>\n"
-                    )
-                    player_user_message += "\n"
-
-            # Create player's prompt in chat format
-            player_prompt = [
-                {"role": "system", "content": system_prompt},
-                {"role": "user", "content": player_user_message},
-            ]
-
-            # Repeat player's prompt for each generation
-            for _ in range(self.num_generations):
-                player_prompts.append(player_prompt)
-
-        return player_prompts
-
-    def _format_bob_prompts(
-        self,
-        original_prompts_batch: List[List[str]],
-        reference_solutions_batch: List[List[str]],
-    ) -> List[List[Dict[str, Any]]]:
-        """
-        Format prompts for Bob given the original prompts and reference solutions.
-
-        Bob identifies removable properties from the reference solutions.
-
-        Args:
-            original_prompts_batch: List of original prompt messages (chat format)
-            reference_solutions_batch: List of reference solution code strings
-
-        Returns:
-            List of formatted prompts for Bob (chat format)
-        """
-        return self._format_player_prompts(
-            original_prompts_batch, reference_solutions_batch, self.bob_prompt
-        )
-
-    def _format_charlie_prompts(
-        self,
-        original_prompts_batch: List[List[str]],
-        reference_solutions_batch: List[List[str]],
-    ) -> List[List[Dict[str, Any]]]:
-        """
-        Format prompts for Charlie given the original prompts and reference solutions.
-
-        Charlie generates alternative solutions that differ from the reference solutions.
-
-        Args:
-            original_prompts_batch: List of original prompt messages (chat format)
-            reference_solutions_batch: List of reference solution code strings
-
-        Returns:
-            List of formatted prompts for Charlie (chat format)
-        """
-        return self._format_player_prompts(
-            original_prompts_batch, reference_solutions_batch, self.charlie_prompt
-        )
-
-    def _set_signature_columns_if_needed(self):
-        # If `self.args.remove_unused_columns` is True, non-signature columns are removed.
-        # By default, this method sets `self._signature_columns` to the model's expected inputs.
-        # In GRPOTrainer, we preprocess data, so using the model's signature columns doesn't work.
-        # Instead, we set them to the columns expected by the `training_step` method, hence the override.
-        if self._signature_columns is None:
-            self._signature_columns = ["prompt", "image", "images"]
-
-    # This method overrides `Trainer.get_train_dataloader` to support our custom batching strategy.
-    # Instead of returning a standard per-step batch (i.e., `per_device_batch_size), our dataloader loads an
-    # *generation* batch (i.e., `per_device_batch_size × steps_per_generation`). This allows us to generate completions
-    # once every steps_per_generation step—rather than once per accumulation step—which is significantly more
-    # efficient. The only change from the original implementation is multiplying the batch size by
-    # `steps_per_generation`. Thus, `_prepare_inputs` is called with this *generation* batch, and it handles the
-    # splitting internally.
-    # Maintenance note: This method is a copy-paste of the original `Trainer.get_train_dataloader` with only one line
-    # modification. As a result, some parts of the method aren't relevant to GRPO, but we keep them to stay one line
-    # apart from the super method, ensuring easier maintenance in the future.
-    def get_train_dataloader(self):
-        if self.train_dataset is None:
-            raise ValueError("Trainer: training requires a train_dataset.")
-
-        train_dataset = self.train_dataset
-        data_collator = self.data_collator
-        if is_datasets_available() and isinstance(train_dataset, datasets.Dataset):
-            train_dataset = self._remove_unused_columns(
-                train_dataset, description="training"
-            )
-        else:
-            data_collator = self._get_collator_with_removed_columns(
-                data_collator, description="training"
-            )
-
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        # Bob and Charlie each need num_references, so we need 2 * num_references total
-        num_ref_samples = (
-            (self.args.generation_batch_size // self.num_generations)
-            * 2
-            * self.num_references
-        )
-
-        dataloader_params = {
-            "batch_size": self._train_batch_size * self.args.steps_per_generation
-            + num_ref_samples,  # < this is the change
-            ############### END GRPOGAMETRAINER CHANGE ###############
-            "collate_fn": data_collator,
-            "num_workers": self.args.dataloader_num_workers,
-            "pin_memory": self.args.dataloader_pin_memory,
-            "persistent_workers": self.args.dataloader_persistent_workers,
-        }
-
-        if not isinstance(train_dataset, torch.utils.data.IterableDataset):
-            dataloader_params["sampler"] = self._get_train_sampler()
-            dataloader_params["drop_last"] = self.args.dataloader_drop_last
-            dataloader_params["worker_init_fn"] = partial(
-                seed_worker,
-                num_workers=self.args.dataloader_num_workers,
-                rank=self.args.process_index,
-            )
-
-            dataloader_params["prefetch_factor"] = self.args.dataloader_prefetch_factor
-
-        return self.accelerator.prepare(DataLoader(train_dataset, **dataloader_params))
-
-    def _get_train_sampler(self, dataset: Optional[Dataset] = None) -> Sampler:
-        """
-        Override to create (n+2*num_references) samples per unique prompt.
-
-        The user configures num_generations=n, but we need to generate n+2*num_references
-        completions (n training + num_references for Bob + num_references for Charlie).
-        The RepeatSampler will repeat each unique sample (n+2*num_references) times.
-        """
-        # Returns a sampler that
-        # 1. ensures each prompt is repeated across multiple processes. This guarantees that identical prompts are
-        #    distributed to different GPUs, allowing rewards to be computed and normalized correctly within each prompt
-        #    group. Using the same seed across processes ensures consistent prompt assignment, preventing discrepancies
-        #    in group formation.
-        # 2. repeats the batch multiple times to allow reusing generations across multiple updates. Refer to
-        #    _prepare_inputs to see how the generations are stored and reused.
-
-        # In the following figure, the values are the prompt indices. The first row shows the first sampled batch, the
-        # second row shows the second sampled batch, and so on.
-        #
-        #                                      |   GPU 0  |   GPU 1  |
-        #
-        #                 global_step   step    <-───>  num_generations=2
-        #                                       <-───────> per_device_train_batch_size=3
-        #  grad_accum    ▲  ▲  0          0     0   0   1   1   2   2   <- Generate for the first `steps_per_generation` (prompts 0 to 11); store the completions; use the first slice to compute the loss
-        #     =2         ▼  |  0          1     3   3   4   4   5   5   <- Take the stored generations and use the second slice to compute the loss
-        #                   |
-        #                   |  1          2     6   6   7   7   8   8   <- Take the stored generations and use the third slice to compute the loss
-        #  steps_per_gen=4  ▼  1          3     9   9  10  10  11  11   <- Take the stored generations and use the fourth slice to compute the loss
-        #
-        #                      2          4    12  12  13  13  14  14   <- Generate for the second `steps_per_generation` (prompts 12 to 23); store the completions; use the first slice to compute the loss
-        #                      2          5    15  15  16  16  17  17   <- Take the stored generations and use the second slice to compute the loss
-        #                                          ...
-        if dataset is None:
-            dataset = self.train_dataset
-        return RepeatSampler(
-            data_source=dataset,
-            mini_repeat_count=self.num_generations + 2 * self.num_references,
-            batch_size=self.args.generation_batch_size // self.num_generations,
-            repeat_count=self.num_iterations * self.args.steps_per_generation,
-            shuffle=self.shuffle_dataset,
-            seed=self.args.seed,
-        )
-
-    def _get_eval_sampler(self, eval_dataset) -> Sampler:
-        # See _get_train_sampler for an explanation of the sampler.
-        num_generations = self.num_generations_eval or self.num_generations
-        return RepeatSampler(
-            data_source=eval_dataset,
-            mini_repeat_count=num_generations,
-            seed=self.args.seed,
-        )
-
-    @profiling_decorator
-    def _get_last_hidden_state(
-        self,
-        unwrapped_model,
-        input_ids,
-        attention_mask,
-        logits_to_keep,
-        pixel_values=None,
-        image_grid_thw=None,
-        pixel_attention_mask=None,
-        image_sizes=None,
-    ):
-        if is_peft_model(unwrapped_model):
-            unwrapped_model = unwrapped_model.base_model.model
-
-        # Build model inputs - check if the model supports logits_to_keep (some models and VLMs don't)
-        model_inputs = {"input_ids": input_ids, "attention_mask": attention_mask}
-
-        # For Qwen models:
-        if image_grid_thw is not None and pixel_values is not None:
-            model_inputs["image_grid_thw"] = image_grid_thw
-        # For Gemma, SmolVLM2, LLaVa-Next etc.:
-        if pixel_values is not None:
-            model_inputs["pixel_values"] = pixel_values
-        # For SmolVLM2
-        if pixel_attention_mask is not None:
-            model_inputs["pixel_attention_mask"] = pixel_attention_mask
-        # For LLaVa-Next
-        if image_sizes is not None:
-            model_inputs["image_sizes"] = image_sizes
-
-        # Only add logits_to_keep if the model supports it
-        if "logits_to_keep" in self.model_kwarg_keys:
-            # We add 1 to `logits_to_keep` because the last logits of the sequence is later excluded
-            model_inputs["logits_to_keep"] = logits_to_keep + 1
-
-        model_inputs["use_cache"] = (
-            False  # only used in generation; set False to suppress warnings
-        )
-
-        last_hidden_state = unwrapped_model.model(**model_inputs).last_hidden_state
-        # Exclude the last value: it corresponds to the next token pred
-        last_hidden_state = last_hidden_state[:, :-1, :]  # (B, L-1, H)
-        # Only keep the last logits_to_keep. For model that support logits_to_keep, this is a no-op.
-        last_hidden_state = last_hidden_state[
-            :, -logits_to_keep:, :
-        ]  # (B, logits_to_keep, H)
-        return last_hidden_state
-
-    def get_high_entropy_mask(
-        self, entropies: torch.Tensor, mask: torch.Tensor, threshold: float
-    ) -> torch.Tensor:
-        """
-        Returns a binary mask identifying tokens whose entropy exceeds a given quantile threshold.
-
-        Args:
-            entropies (`torch.Tensor`):
-                Tensor of shape (batch_size, seq_len) with per-token entropy values.
-            mask (`torch.Tensor`):
-                Binary mask of the same shape as `entropies`, where `1` indicates valid tokens and `0` padding.
-            threshold (`float`):
-                Quantile threshold between `0.0` and `1.0` to select high-entropy tokens.
-
-        Returns:
-            `torch.Tensor`:
-                Boolean mask of shape (batch_size, seq_len), where `True` indicates tokens with entropy >= threshold
-                and `False` otherwise.
-        """
-        local = entropies[mask.bool()].float()
-
-        # Use a negative pad_value as a sentinel because entropy values are always >= 0.
-        # This guarantees that the sentinel cannot collide with any real entropy value.
-        pad_value = -1e9
-
-        # Pad across processes so that every rank has the same tensor length
-        padded = self.accelerator.pad_across_processes(
-            local, dim=0, pad_index=pad_value
-        )
-        gathered = self.accelerator.gather(padded)
-
-        # Drop sentinel values (safe because no entropy can be negative)
-        gathered = gathered[gathered != pad_value]
-
-        if gathered.numel() == 0:
-            return torch.zeros_like(entropies, dtype=torch.bool)
-
-        entropy_threshold = torch.quantile(gathered, threshold)
-        masked_entropies = entropies * mask.float()
-        entropy_mask = masked_entropies >= entropy_threshold
-        return entropy_mask & mask.bool()  # ensure padding tokens are always masked out
-
-    @profiling_decorator
-    def _get_per_token_logps_and_entropies(
-        self,
-        model,
-        input_ids,
-        attention_mask,
-        logits_to_keep,
-        batch_size=None,
-        compute_entropy=False,
-        pixel_values=None,
-        image_grid_thw=None,
-        num_images=None,
-        pixel_attention_mask=None,
-        image_sizes=None,
-        token_type_ids=None,
-    ) -> dict[str, Optional[torch.Tensor]]:
-        """Compute log-probs and (optionally) entropies for each token."""
-        batch_size = batch_size or input_ids.size(
-            0
-        )  # Chunk inputs into smaller batches to reduce memory peak
-        all_logps = []
-        all_entropies = []
-        for start in range(0, input_ids.size(0), batch_size):
-            input_ids_batch = input_ids[start : start + batch_size]
-            attention_mask_batch = attention_mask[start : start + batch_size]
-
-            # Build model inputs - check if the model supports logits_to_keep (some models and VLMs don't)
-            model_inputs = {
-                "input_ids": input_ids_batch,
-                "attention_mask": attention_mask_batch,
-            }
-            if image_grid_thw is not None and pixel_values is not None:
-                rows_per_image = image_grid_thw.prod(dim=-1)
-                rows_per_sample = torch.split(rows_per_image, num_images)
-                rows_per_sample = torch.stack([s.sum() for s in rows_per_sample])
-                cum_rows = torch.cat(
-                    [
-                        torch.tensor([0], device=rows_per_sample.device),
-                        rows_per_sample.cumsum(0),
-                    ]
-                )
-                row_start, row_end = (
-                    cum_rows[start].item(),
-                    cum_rows[start + batch_size].item(),
-                )
-                model_inputs["pixel_values"] = pixel_values[row_start:row_end]
-                cum_imgs = torch.tensor([0] + num_images).cumsum(0)
-                img_start, img_end = cum_imgs[start], cum_imgs[start + batch_size]
-                model_inputs["image_grid_thw"] = image_grid_thw[img_start:img_end]
-            elif pixel_values is not None:
-                model_inputs["pixel_values"] = pixel_values[start : start + batch_size]
-            if pixel_attention_mask is not None:
-                model_inputs["pixel_attention_mask"] = pixel_attention_mask[
-                    start : start + batch_size
-                ]
-            if image_sizes is not None:
-                model_inputs["image_sizes"] = image_sizes[start : start + batch_size]
-            if token_type_ids is not None:
-                model_inputs["token_type_ids"] = token_type_ids[
-                    start : start + batch_size
-                ]
-
-            # Only add logits_to_keep if the model supports it
-            if "logits_to_keep" in self.model_kwarg_keys:
-                # We add 1 to `logits_to_keep` because the last logits of the sequence is later excluded
-                model_inputs["logits_to_keep"] = logits_to_keep + 1
-
-            model_inputs["use_cache"] = (
-                False  # only used in generation; set False to suppress warnings
-            )
-
-            logits = model(**model_inputs).logits
-            # Exclude the last value: it corresponds to the next token pred
-            logits = logits[:, :-1, :]  # (B, L-1, H)
-            # Only keep the last logits_to_keep. For model that support logits_to_keep, this is a no-op.
-            logits = logits[:, -logits_to_keep:, :]  # (B, logits_to_keep, H)
-            # Divide logits by sampling temperature.
-            # See https://huggingface.co/blog/the_n_implementation_details_of_rlhf_with_ppo#policy-training-implementation-details
-            logits = logits / self.temperature
-
-            completion_ids = input_ids_batch[:, -logits_to_keep:]
-            logps = selective_log_softmax(logits, completion_ids)  # compute logprobs
-            all_logps.append(logps)
-
-            if compute_entropy:
-                with torch.no_grad():
-                    entropies = entropy_from_logits(logits)
-                all_entropies.append(entropies)
-
-        logps = torch.cat(all_logps, dim=0)
-        entropies = torch.cat(all_entropies, dim=0) if compute_entropy else None
-        return logps, entropies
-
-    def _fix_param_name_to_vllm(self, name, extra_prefixes: Optional[list[str]] = None):
-        extra_prefixes = extra_prefixes or []
-        prefixes = ["_checkpoint_wrapped_module."] + extra_prefixes
-        for prefix in prefixes:
-            name = name.replace(prefix, "")
-        return name
-
-    def _sync_fsdp1_params_to_vllm(
-        self, module: nn.Module, prefix: str = "", visited=None
-    ):
-        """Memory-efficient post-order traversal of FSDP modules to extract full parameters and sync with vLLM."""
-        # For FSDP1, we need to recurse into children and also use summon_full_params
-        if visited is None:
-            visited = set()
-        for child_name, child_module in module.named_children():
-            child_prefix = f"{prefix}.{child_name}" if prefix else child_name
-            self._sync_fsdp1_params_to_vllm(
-                child_module, prefix=child_prefix, visited=visited
-            )  # recurse into the child
-
-        if isinstance(module, FSDP):
-            with FSDP.summon_full_params(module, recurse=False, writeback=False):
-                for param_name, param in module.named_parameters():
-                    full_name = f"{prefix}.{param_name}" if prefix else param_name
-                    full_name = self._fix_param_name_to_vllm(
-                        full_name, extra_prefixes=["_fsdp_wrapped_module."]
-                    )
-
-                    if full_name in visited:
-                        continue  # skip FSDP subtrees already traversed
-                    visited.add(full_name)
-
-                    if self.vllm_mode == "server" and self.accelerator.is_main_process:
-                        self.vllm_client.update_named_param(full_name, param.data)
-                    elif self.vllm_mode == "colocate":
-                        llm_model = self.llm.llm_engine.model_executor.driver_worker.model_runner.model
-                        llm_model.load_weights([(full_name, param.data)])
-
-    def _sync_fsdp2_params_to_vllm(self, module: nn.Module):
-        # For FSDP2, module.state_dict() already covers all parameters, so no need for recursion
-        for name, param in module.state_dict().items():
-            if param.is_cpu:
-                param = param.to(torch.device("cuda"))
-            param = param.full_tensor()
-
-            if self.vllm_mode == "server" and self.accelerator.is_main_process:
-                self.vllm_client.update_named_param(name, param)
-            elif self.vllm_mode == "colocate":
-                llm_model = (
-                    self.llm.llm_engine.model_executor.driver_worker.model_runner.model
-                )
-                llm_model.load_weights([(name, param)])
-
-    @profiling_decorator
-    def _move_model_to_vllm(self):
-        # For DeepSpeed ZeRO-3 and FSDP, we need to gather all parameters before operations
-        deepspeed_plugin = self.accelerator.state.deepspeed_plugin
-        zero_stage_3 = deepspeed_plugin is not None and deepspeed_plugin.zero_stage == 3
-        if zero_stage_3:
-            import deepspeed
-
-            gather_if_zero3 = deepspeed.zero.GatheredParameters
-        else:
-            gather_if_zero3 = nullcontext
-
-        if is_peft_model(self.model):
-            # With PEFT and FSDP/DeepSpeed ZeRO Stage 3, we must gather the full model at once before merging, as
-            # merging adapters in a sharded manner is not supported.
-            # TODO: does this work with FSDP?
-            with gather_if_zero3(list(self.model.parameters())):
-                self.model.merge_adapter()
-
-                # Update vLLM weights while parameters are gathered
-                if (
-                    self.is_fsdp_enabled
-                ):  # note if using FSDP, gather_if_zero3 is nullcontext
-                    # Update vLLM weights while parameters are gathered
-                    # For PEFT with FSDP we need to use the memory efficient post-order traversal
-                    fsdp_plugin = getattr(self.accelerator.state, "fsdp_plugin", None)
-                    fsdp_version = (
-                        getattr(fsdp_plugin, "fsdp_version", 1) if fsdp_plugin else 1
-                    )
-                    if fsdp_version == 1:
-                        self._sync_fsdp1_params_to_vllm(
-                            self.model
-                        )  # use memory-efficient post-order traversal for FSDP
-                    elif fsdp_version == 2:
-                        self._sync_fsdp2_params_to_vllm(self.model)
-                else:
-                    # DeepSpeed ZeRO-3 with PEFT
-                    for name, param in self.model.named_parameters():
-                        # When using PEFT, we need to recover the original parameter name and discard some parameters
-                        name = name.removeprefix("base_model.model.").replace(
-                            ".base_layer", ""
-                        )
-                        if self.model.prefix in name:
-                            continue
-                        # When module to save, remove its prefix and discard the original module
-                        if "original_module" in name:
-                            continue
-                        name = self._fix_param_name_to_vllm(
-                            name, extra_prefixes=["modules_to_save.default."]
-                        )
-
-                        if (
-                            self.vllm_mode == "server"
-                            and self.accelerator.is_main_process
-                        ):
-                            self.vllm_client.update_named_param(name, param.data)
-                        elif self.vllm_mode == "colocate":
-                            llm_model = self.llm.llm_engine.model_executor.driver_worker.model_runner.model
-                            llm_model.load_weights([(name, param.data)])
-                # Unmerge adapters while parameters are still gathered
-                self.model.unmerge_adapter()
-                # Parameters will automatically be repartitioned when exiting the context
-        else:
-            # For non-PEFT models, simply gather (if needed) and update each parameter individually.
-            if self.is_fsdp_enabled:
-                fsdp_plugin = getattr(self.accelerator.state, "fsdp_plugin", None)
-                fsdp_version = (
-                    getattr(fsdp_plugin, "fsdp_version", 1) if fsdp_plugin else 1
-                )
-                if fsdp_version == 1:
-                    self._sync_fsdp1_params_to_vllm(
-                        self.model
-                    )  # use memory-efficient post-order traversal for FSDP
-                elif fsdp_version == 2:
-                    self._sync_fsdp2_params_to_vllm(self.model)
-            else:
-                for name, param in self.model.named_parameters():
-                    name = self._fix_param_name_to_vllm(name)
-                    with gather_if_zero3([param]):
-                        if (
-                            self.vllm_mode == "server"
-                            and self.accelerator.is_main_process
-                        ):
-                            self.vllm_client.update_named_param(name, param.data)
-                        elif self.vllm_mode == "colocate":
-                            llm_model = self.llm.llm_engine.model_executor.driver_worker.model_runner.model
-                            llm_model.load_weights([(name, param.data)])
-
-        # Reset cache on vLLM
-        if self.vllm_mode == "server" and self.accelerator.is_main_process:
-            self.vllm_client.reset_prefix_cache()
-        elif self.vllm_mode == "colocate":
-            self.llm.reset_prefix_cache()
-
-    @profiling_decorator
-    def _prepare_inputs(
-        self, generation_batch: dict[str, Union[torch.Tensor, Any]]
-    ) -> dict[str, Union[torch.Tensor, Any]]:
-        # Prepares inputs for model training/evaluation by managing completion generation and batch handling.
-        # During training:
-        #   - Receives the local generation batch (Per-GPU batch size × steps per generation)
-        #     from the modified training dataloader instead of the standard local batch
-        #   - Generates completions once for the entire generation batch and splits it into batches of size
-        #     `per_device_train_batch_size`
-        #   - Buffers these completions and returns the appropriate slice for the current accumulation step
-        #   - Optimizes by regenerating completions only periodically (every steps_per_generation * num_iterations)
-        # During evaluation:
-        #   - The input is treated as a standard local batch (no accumulation, no multiple iterations)
-        #   - Completions are generated for each batch without buffering or reuse
-        # Returns a single local batch in both cases.
-
-        mode = "train" if self.model.training else "eval"
-        if mode == "train":
-            generate_every = self.args.steps_per_generation * self.num_iterations
-            if self._step % generate_every == 0 or self._buffered_inputs is None:
-                # self._buffered_inputs=None can occur when resuming from a checkpoint
-                generation_batch = self._generate_and_score_completions(
-                    generation_batch
-                )
-                generation_batch = split_pixel_values_by_grid(generation_batch)
-                generation_batch = shuffle_sequence_dict(generation_batch)
-                generation_batches = split_tensor_dict(
-                    generation_batch, self.args.steps_per_generation
-                )
-                self._buffered_inputs = [
-                    unsplit_pixel_values_by_grid(batch) for batch in generation_batches
-                ]
-            inputs = self._buffered_inputs[self._step % self.args.steps_per_generation]
-            self._step += 1
-        else:
-            # In evaluation, there is neither batch grouping for generation, nor multiple iterations, hence
-            # local generation batch == local eval batch
-            inputs = self._generate_and_score_completions(generation_batch)
-        return inputs
-
-    @profiling_decorator
-    def _calculate_rewards(self, inputs, prompts, completions, completion_ids_list):
-        device = self.accelerator.device
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if not self.model.training:
-            rewards_per_func = torch.zeros(
-                len(prompts), len(self.reward_funcs), device=device
-            )
-        else:
-            rewards_per_func = torch.zeros(
-                # Bob completions are appended to the training completions,
-                # so we need to multiply the number of prompts by 2
-                len(prompts) * 2,
-                len(self.reward_funcs) + 1,  # +1 for the game reward
-                device=device,
-            )
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        # Repeat all input columns (but "prompt", "completion", and "completion_ids") to match the num of generations
-        keys = [
-            key
-            for key in inputs[0]
-            if key not in ["prompt", "completion", "completion_ids"]
-        ]
-        reward_kwargs = {key: [example[key] for example in inputs] for key in keys}
-
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            # Append Charlie's solutions to the training completions for scoring
-            # (Bob only generates properties, Charlie generates alternative solutions)
-            prompts.extend(
-                self._charlie_prompts
-            )  # We want to compute rewards as if they came from the same prompts
-            completions.extend(self._charlie_solutions)
-            completion_ids_list.extend(self._charlie_solutions_completion_ids)
-
-            # Extend reward_kwargs to match the doubled length
-            # (Alice's + Charlie's data)
-            for key in keys:
-                reward_kwargs[key].extend(reward_kwargs[key])
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        # This allows for dynamic reward shaping based on training progress.
-        reward_kwargs["trainer_state"] = self.state
-
-        for i, (reward_func, reward_processing_class, reward_func_name) in enumerate(
-            zip(
-                self.reward_funcs,
-                self.reward_processing_classes,
-                self.reward_func_names,
-            )
-        ):
-            with profiling_context(self, reward_func_name):
-                if isinstance(
-                    reward_func, nn.Module
-                ):  # Module (no PretrainedModel) for compat with compiled models
-                    if is_conversational(inputs[0]):
-                        messages = [
-                            {"messages": p + c} for p, c in zip(prompts, completions)
-                        ]
-                        texts = [
-                            apply_chat_template(x, reward_processing_class)["text"]
-                            for x in messages
-                        ]
-                    else:
-                        texts = [p + c for p, c in zip(prompts, completions)]
-                    reward_inputs = reward_processing_class(
-                        text=texts,
-                        return_tensors="pt",
-                        padding=True,
-                        padding_side="right",
-                        add_special_tokens=False,
-                    )
-                    reward_inputs = super()._prepare_inputs(reward_inputs)
-                    with torch.inference_mode():
-                        rewards_per_func[:, i] = reward_func(**reward_inputs).logits[
-                            :, 0
-                        ]  # Shape (B*G,)
-                else:
-                    output_reward_func = reward_func(
-                        prompts=prompts,
-                        completions=completions,
-                        completion_ids=completion_ids_list,
-                        **reward_kwargs,
-                    )
-                    # Convert None values to NaN
-                    output_reward_func = [
-                        reward if reward is not None else torch.nan
-                        for reward in output_reward_func
-                    ]
-
-                    rewards_per_func[:, i] = torch.tensor(
-                        output_reward_func, dtype=torch.float32, device=device
-                    )
-
-        # If all reward functions return None for a given row, issue a detailed warning
-        if torch.isnan(rewards_per_func).all(dim=1).any():
-            nan_row_idx = (
-                torch.isnan(rewards_per_func).all(dim=1).nonzero(as_tuple=True)[0][0]
-            )
-            row_reward_kwargs = {
-                key: value[nan_row_idx]
-                for key, value in reward_kwargs.items()
-                if key != "trainer_state"
-            }
-            row_reward_kwargs["prompt"] = prompts[nan_row_idx]
-            row_reward_kwargs["completion"] = completions[nan_row_idx]
-            logger.warning(
-                f"All reward functions returned None for the following kwargs:\n{row_reward_kwargs}\n"
-                "Please ensure that at least one reward function returns a valid reward."
-            )
-
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            # Split Alice's and Charlie's rewards into separate tensors.
-            # Note: Bob generates properties, Charlie generates alternative solutions
-            num_alice = len(prompts) // 2  # Half are Alice's, half are Charlie's
-            alice_rewards = rewards_per_func[:num_alice, :]
-            charlie_rewards = rewards_per_func[num_alice:, :]
-
-            # Extract Alice's and Charlie's solutions
-            alice_completions = [completions[i] for i in range(num_alice)]
-            charlie_completions = [
-                completions[i] for i in range(num_alice, len(completions))
-            ]
-
-            alice_solutions = []
-            for completion in alice_completions:
-                if isinstance(completion, list) and len(completion) > 0:
-                    if isinstance(completion[0], dict) and "content" in completion[0]:
-                        alice_solutions.append(completion[0]["content"])
-                    else:
-                        alice_solutions.append(str(completion))
-                else:
-                    alice_solutions.append(str(completion))
-
-            charlie_solutions_text = []
-            for solution in charlie_completions:
-                if isinstance(solution, list) and len(solution) > 0:
-                    if isinstance(solution[0], dict) and "content" in solution[0]:
-                        charlie_solutions_text.append(solution[0]["content"])
-                    else:
-                        charlie_solutions_text.append(str(solution))
-                else:
-                    charlie_solutions_text.append(str(solution))
-
-            solutions_to_check = alice_solutions + charlie_solutions_text
-            # Repeat properties (from Bob) for batched checking
-            properties_to_check = self._bob_properties * 2
-            if self.solution_preprocessor is not None:
-                solutions_to_check = [
-                    self.solution_preprocessor(solution)
-                    for solution in solutions_to_check
-                ]
-
-            # Check properties for Alice's and Charlie's solutions
-            has_properties = self._check_properties(
-                solutions_to_check, properties_to_check
-            )
-            has_properties_alice = has_properties[: len(alice_solutions)]
-            has_properties_charlie = has_properties[len(alice_solutions) :]
-
-            # Extract scores from rewards (sum across all reward functions)
-            # Convert to float lists for game computation
-            # Ensure game_score_weights is on the same device as the rewards
-            game_score_weights = self.game_score_weights.to(alice_rewards.device)
-            alice_scores = (
-                (alice_rewards * game_score_weights).sum(dim=1).cpu().tolist()
-            )
-            charlie_scores = (
-                (charlie_rewards * game_score_weights).sum(dim=1).cpu().tolist()
-            )
-
-            # Compute game rewards
-            # Bob and Charlie are on the same team, so Charlie's solutions compete with Alice's
-            self._game_rewards = self._get_game_rewards(
-                alice_solutions=alice_solutions,
-                bob_solutions=charlie_solutions_text,  # Using Charlie's solutions
-                properties=self._bob_properties,
-                has_properties_alice=has_properties_alice,
-                has_properties_bob=has_properties_charlie,  # Charlie's property flags
-                alice_scores=alice_scores,
-                bob_scores=charlie_scores,  # Charlie's scores
-            )
-
-            if is_wandb_available():
-                table = wandb.Table(
-                    columns=[
-                        "question",
-                        "alice_completion",
-                        "property",
-                        "charlie_property",
-                        "alice_flagged",
-                        "charlie_completion",
-                        "charlie_flagged",
-                        "alice_score",
-                        "charlie_score",
-                        "Reward",
-                    ]
-                )
-                for i in range(len(prompts) // 2):
-                    table.add_data(
-                        prompts[i][1]["content"],
-                        alice_solutions[i],
-                        self._bob_properties[i],
-                        self._charlie_properties[i],
-                        has_properties_alice[i],
-                        charlie_solutions_text[i],
-                        has_properties_charlie[i],
-                        alice_scores[i],
-                        charlie_scores[i],
-                        self._game_rewards[i],
-                    )
-                try:
-                    wandb.log({"game_table": table})
-                except wandb.errors.AuthenticationError:  # type: ignore
-                    print("Authentication error when logging game table. Skipping.")
-
-            alice_rewards[:, -1] = torch.tensor(
-                self._game_rewards, device=device, dtype=torch.float32
-            )
-            rewards_per_func = gather(alice_rewards)
-            return rewards_per_func
-        ############### END GRPOGAMETRAINER CHANGE ###############
-        # Gather Alice's rewards: this part is crucial, because the rewards are normalized per group and the
-        # completions may be distributed across processes
-        rewards_per_func = gather(rewards_per_func)
-
-        return rewards_per_func
-
-    ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-    def _generate_single_turn(self, prompts: list[str], images: Optional[list]):
-        """
-        Override to generate n+1 completions per prompt instead of n.
-
-        The sampler provides (n+1) * num_unique_prompts duplicates.
-        We temporarily increase self.num_generations to n+1, call parent's
-        generation, then restore it. The reference completion
-        (last one in each group) will be extracted and stored.
-        """
-        # During evaluation, skip all game mechanisms and use core implementation
-        if not self.model.training:
-            prompt_ids, completion_ids, logprobs, forward_kwargs = (
-                self._generate_single_turn_core(prompts, images)
-            )
-
-            return (
-                prompt_ids,
-                completion_ids,
-                logprobs,
-                forward_kwargs,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-            )
-
-        # Store original num_generations
-        original_num_gens = self.num_generations
-
-        try:
-            # Temporarily set to n + 2*num_references for generation
-            # Bob and Charlie will each see different reference solutions
-            self.num_generations = original_num_gens + 2 * self.num_references
-            prompt_ids, completion_ids, logprobs, forward_kwargs = (
-                self._generate_single_turn_core(prompts, images)
-            )
-
-            # Extract reference completions into list of lists
-            # Structure: batch_size * (original_num_gens + 2*num_references) total completions
-            # Each batch item has:
-            #   - original_num_gens regular completions (for Alice)
-            #   - num_references reference completions (for Bob)
-            #   - num_references reference completions (for Charlie)
-            bob_reference_completion_ids = [
-                completion_ids[
-                    batch_start + original_num_gens : batch_start
-                    + original_num_gens
-                    + self.num_references
-                ]
-                for batch_start in range(0, len(completion_ids), self.num_generations)
-            ]
-            bob_reference_prompts = [
-                prompts[
-                    batch_start + original_num_gens : batch_start
-                    + original_num_gens
-                    + self.num_references
-                ]
-                for batch_start in range(0, len(completion_ids), self.num_generations)
-            ]
-
-            charlie_reference_completion_ids = [
-                completion_ids[
-                    batch_start + original_num_gens + self.num_references : batch_start
-                    + original_num_gens
-                    + 2 * self.num_references
-                ]
-                for batch_start in range(0, len(completion_ids), self.num_generations)
-            ]
-            charlie_reference_prompts = [
-                prompts[
-                    batch_start + original_num_gens + self.num_references : batch_start
-                    + original_num_gens
-                    + 2 * self.num_references
-                ]
-                for batch_start in range(0, len(completion_ids), self.num_generations)
-            ]
-
-            self.num_generations = original_num_gens
-
-            # Generate Bob's completions (properties only) using Bob's references
-            (
-                bob_prompt_ids,
-                bob_completion_ids,
-                bob_logprobs,
-                bob_forward_kwargs,
-            ) = self._generate_bob_completions(
-                bob_reference_completion_ids, bob_reference_prompts, images
-            )
-
-            # Generate Charlie's completions (alternative solutions only) using Charlie's references
-            (
-                charlie_prompt_ids,
-                charlie_completion_ids,
-                charlie_logprobs,
-                charlie_forward_kwargs,
-            ) = self._generate_charlie_completions(
-                charlie_reference_completion_ids, charlie_reference_prompts, images
-            )
-
-            return (
-                prompt_ids,
-                completion_ids,
-                logprobs,
-                forward_kwargs,
-                bob_prompt_ids,
-                bob_completion_ids,
-                bob_logprobs,
-                bob_forward_kwargs,
-                charlie_prompt_ids,
-                charlie_completion_ids,
-                charlie_logprobs,
-                charlie_forward_kwargs,
-            )
-        finally:
-            self.num_generations = original_num_gens
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-    def _generate_single_turn_core(self, prompts: list[str], images: Optional[list]):
-        device = self.accelerator.device
-        mode = "train" if self.model.training else "eval"
-        num_generations = (
-            self.num_generations
-            if mode == "train"
-            else self.num_generations_eval or self.num_generations
-        )
-
-        # If the prompts are conversational and the inputs contain images, we need to convert the prompts from
-        # [{"role": "user", "content": "What color is the sky?"}] to
-        # [{"role": "user", "content": [{"type": "image"}, {"type": "text", "text": "What color is the sky?"}]}]
-        kwargs = {}
-        if images is not None:
-            kwargs = {"images": images}
-            for prompt, image_list in zip(prompts, images):
-                if isinstance(prompt, list):  # i.e., when using conversational data
-                    prepare_multimodal_messages(prompt, num_images=len(image_list))
-
-        prompts_text = [
-            maybe_apply_chat_template({"prompt": prompt}, self.processing_class)[
-                "prompt"
-            ]
-            for prompt in prompts
-        ]
-
-        prompt_inputs = self.processing_class(
-            text=prompts_text,
-            return_tensors="pt",
-            padding=True,
-            padding_side="left",
-            add_special_tokens=False,
-            **kwargs,
-        )
-        prompt_inputs = super()._prepare_inputs(prompt_inputs)
-        prompt_ids, prompt_mask = (
-            prompt_inputs["input_ids"],
-            prompt_inputs["attention_mask"],
-        )
-        forward_kwargs = {
-            k: v
-            for k, v in prompt_inputs.items()
-            if k not in ["input_ids", "attention_mask"]
-        }
-        prompt_ids = [p[m].tolist() for p, m in zip(prompt_ids, prompt_mask.bool())]
-
-        if self.max_prompt_length is not None:
-            # If max_prompt_length is set, we trim the prompt to keep only the last `max_prompt_length` tokens.
-            # Then we decode those tokens back into text. We set `skip_special_tokens=False` because some special
-            # tokens are needed for generation.
-            protected = [
-                self.image_token_id,
-                self.vision_start_token_id,
-                self.vision_end_token_id,
-            ]
-            protected = [token for token in protected if token is not None]
-            prompt_ids = [
-                truncate_with_protected_tokens(ids, self.max_prompt_length, protected)
-                for ids in prompt_ids
-            ]
-
-            prompts_text = self.processing_class.batch_decode(
-                prompt_ids,
-                skip_special_tokens=False,
-                clean_up_tokenization_spaces=False,
-            )
-
-            # The chat template sometimes inserts a single image token into the prompt text. However, when this text is
-            # later tokenized, the single image token string is expanded into multiple image token IDs, depending on the
-            # image size. Since we're detokenizing here, we may see repeated image tokens in the decoded text. We
-            # collapse them back into a single token string to match the original chat template in case it originally
-            # applies it. Otherwise, it assumes that the chat template uses only vision_start_token_id to indicate images
-            # (e.g. Gemma 3) and removes all image_token instances and vision_end_token_id as well, leaving only
-            # the vision_start_token_id (e.g. <start_of_image>).
-            if self.image_token is not None:
-                escaped_img_token = re.escape(self.image_token)
-                # Search for the image token in the chat template
-                if re.search(escaped_img_token, self.processing_class.chat_template):
-                    prompts_text = [
-                        re.sub(rf"({escaped_img_token})+", self.image_token, text)
-                        for text in prompts_text
-                    ]
-                else:
-                    # If the chat template doesn't use the image token, we remove all instances of it + vision_end_token_id
-                    if self.vision_end_token_id is not None:
-                        escaped_eoi_token = re.escape(
-                            self.processing_class.tokenizer.decode(
-                                [self.vision_end_token_id]
-                            )
-                        )
-                        prompts_text = [
-                            re.sub(
-                                rf"({escaped_img_token})+{escaped_eoi_token}", "", text
-                            )
-                            for text in prompts_text
-                        ]
-                    else:
-                        # If vision_end_token_id is None, just remove the image tokens
-                        prompts_text = [
-                            re.sub(rf"({escaped_img_token})+", "", text)
-                            for text in prompts_text
-                        ]
-
-        # Generate completions using either vLLM or regular generation
-        if self.use_vllm:
-            if self.vllm_mode == "colocate" and self.args.vllm_enable_sleep_mode:
-                # wake up colocated vLLM instances if needed
-                torch.cuda.empty_cache()  # required to avoid OOM in some cases
-                self.llm.wake_up()
-
-            # First, update the vLLM weights if needed
-            if self.state.global_step != self._last_loaded_step:
-                self._move_model_to_vllm()
-                self._last_loaded_step = self.state.global_step
-
-            # Generate completions using vLLM: gather all prompts and use them in a single call in the main process
-            if self.vllm_mode == "server":
-                all_prompts_text = gather_object(prompts_text)
-                if images is not None:
-                    all_images = gather_object(images)
-
-                if self.accelerator.is_main_process:
-                    # Since 'prompts' contains 'num_generations' duplicates, we first take unique prompts, and generate
-                    # num_generations outputs for each one. This is faster than generating outputs for each duplicate
-                    # prompt individually.
-                    ordered_set_of_prompts = all_prompts_text[::num_generations]
-
-                    if images is not None:
-                        ordered_set_of_images = all_images[::num_generations]
-                    else:
-                        ordered_set_of_images = None
-
-                    with profiling_context(self, "vLLM.generate"):
-                        output = self.vllm_client.generate(
-                            prompts=ordered_set_of_prompts,
-                            images=ordered_set_of_images,
-                            n=num_generations,
-                            repetition_penalty=self.repetition_penalty,
-                            temperature=self.temperature,
-                            top_p=self.top_p,
-                            top_k=-1 if self.top_k is None else self.top_k,
-                            min_p=0.0 if self.min_p is None else self.min_p,
-                            max_tokens=self.max_completion_length,
-                            guided_decoding_regex=self.guided_decoding_regex,
-                            generation_kwargs=self.args.generation_kwargs,
-                        )
-                        payload = (output["completion_ids"], output["logprobs"])
-                else:
-                    payload = None
-
-                # Broadcast the completions from the main process to all processes, ensuring each process receives its corresponding slice.
-                obj_list = [payload]
-                broadcast_object_list(obj_list, from_process=0)
-                all_completion_ids, all_logprobs = obj_list[0]
-
-                process_slice = slice(
-                    self.accelerator.process_index * len(prompts),
-                    (self.accelerator.process_index + 1) * len(prompts),
-                )
-                completion_ids = all_completion_ids[process_slice]
-                logprobs = all_logprobs[process_slice]
-
-            # Generate completions using colocated vLLM instances: each device holds vLLM copy and work on their own batch of prompts
-            elif self.vllm_mode == "colocate":
-                if self.guided_decoding_regex:
-                    guided_decoding = GuidedDecodingParams(
-                        regex=self.guided_decoding_regex
-                    )
-                else:
-                    guided_decoding = None
-
-                generation_kwargs = {
-                    "n": 1,  # vLLM on each GPU generates only 1 in colocate mode
-                    "repetition_penalty": self.repetition_penalty,
-                    "temperature": self.temperature,
-                    "top_p": self.top_p,
-                    "top_k": -1 if self.top_k is None else self.top_k,
-                    "min_p": 0.0 if self.min_p is None else self.min_p,
-                    "max_tokens": self.max_completion_length,
-                    "guided_decoding": guided_decoding,
-                    "logprobs": 0,  # only return the logprob of the generated token
-                }
-                if self.args.generation_kwargs is not None:
-                    generation_kwargs.update(self.args.generation_kwargs)
-                sampling_params = SamplingParams(**generation_kwargs)
-
-                if self.vllm_tensor_parallel_size > 1:
-                    # Gather prompts from all ranks in the TP group and flatten.
-                    # Each rank starts with its own prompts; after gathering, all ranks see the full group set.
-                    orig_size = len(prompts_text)
-                    gathered_prompts = [
-                        None for _ in range(self.vllm_tensor_parallel_size)
-                    ]
-                    torch.distributed.all_gather_object(
-                        gathered_prompts, prompts_text, group=self.tp_group
-                    )
-                    all_prompts_text = [
-                        p for sublist in gathered_prompts for p in sublist
-                    ]
-
-                    if images is not None:
-                        gathered_images = [
-                            None for _ in range(self.vllm_tensor_parallel_size)
-                        ]
-                        torch.distributed.all_gather_object(
-                            gathered_images, images, group=self.tp_group
-                        )
-                        all_images = [
-                            img for sublist in gathered_images for img in sublist
-                        ]
-                    else:
-                        all_images = None
-                else:
-                    all_prompts_text = prompts_text
-                    all_images = images
-
-                if images is not None and all_images:
-                    vllm_inputs = []
-                    for prompt, image_list in zip(all_prompts_text, all_images):
-                        vllm_inputs.append(
-                            {
-                                "prompt": prompt,
-                                "multi_modal_data": {"image": image_list},
-                            }
-                        )
-
-                else:
-                    vllm_inputs = all_prompts_text
-
-                with profiling_context(self, "vLLM.generate"):
-                    all_outputs = self.llm.generate(
-                        vllm_inputs, sampling_params=sampling_params, use_tqdm=False
-                    )
-
-                all_completion_ids = [
-                    output.token_ids
-                    for outputs in all_outputs
-                    for output in outputs.outputs
-                ]
-                all_logprobs = [
-                    [next(iter(lp.values())).logprob for lp in output.logprobs]
-                    for outputs in all_outputs
-                    for output in outputs.outputs
-                ]
-
-                if self.vllm_tensor_parallel_size > 1:
-                    # Slice completions for this rank within its TP group.
-                    # Each rank generates all outputs — we keep only our share.
-                    local_rank_in_group = torch.distributed.get_rank(
-                        group=self.tp_group
-                    )
-                    tp_slice = slice(
-                        local_rank_in_group * orig_size,
-                        (local_rank_in_group + 1) * orig_size,
-                    )
-                    completion_ids = all_completion_ids[tp_slice]
-                    logprobs = all_logprobs[tp_slice]
-                else:
-                    completion_ids = all_completion_ids
-                    logprobs = all_logprobs
-
-                if self.args.vllm_enable_sleep_mode:
-                    self.llm.sleep(level=1)
-
-        elif self.use_transformers_paged:
-            # Re-process inputs for paged generation if needed
-            # Note: images are already validated and preprocessed above
-            paged_prompt_inputs = self.processing_class(text=prompts_text, **kwargs)
-            previous_attn = self.model_wrapped.config._attn_implementation
-
-            if is_flash_attn_2_available():
-                self.model_wrapped.config._attn_implementation = "paged_attention"
-            else:
-                self.model_wrapped.config._attn_implementation = "sdpa_paged"
-            with (
-                profiling_context(self, "transformers.generate_batch"),
-                unwrap_model_for_generation(
-                    self.model_wrapped,
-                    self.accelerator,
-                    gather_deepspeed3_params=self.args.ds3_gather_for_generation,
-                ) as unwrapped_model,
-                torch.no_grad(),
-                FSDP.summon_full_params(self.model_wrapped, recurse=False)
-                if self.is_fsdp_enabled
-                else nullcontext(),
-            ):
-                # Cast to the appropriate dtype based on training configuration
-                if self.args.bf16:
-                    unwrapped_model.to(torch.bfloat16)
-                elif self.args.fp16:
-                    unwrapped_model.to(torch.float16)
-                with torch.inference_mode():
-                    all_outputs = unwrapped_model.generate_batch(
-                        paged_prompt_inputs.input_ids,
-                        generation_config=self.generation_config,
-                        progress_bar=False,
-                    )
-                    unwrapped_model.train()  # restore training mode, as generate_batch forces eval mode
-            completion_ids = [
-                output.generated_tokens for output in all_outputs.values()
-            ]
-            prompt_ids = paged_prompt_inputs.input_ids
-            # Restore the original attention implementation, training mode
-            self.model_wrapped.config._attn_implementation = previous_attn
-            logprobs = None  # not used in this case
-
-        else:
-            # Regular generation path
-            prompt_ids = [torch.tensor(ids, device=device) for ids in prompt_ids]
-            prompt_mask = [torch.ones_like(ids, dtype=torch.long) for ids in prompt_ids]
-            prompt_ids = pad(
-                prompt_ids, padding_value=self.pad_token_id, padding_side="left"
-            )
-            prompt_mask = pad(prompt_mask, padding_value=0, padding_side="left")
-
-            with (
-                profiling_context(self, "transformers.generate"),
-                unwrap_model_for_generation(
-                    self.model_wrapped,
-                    self.accelerator,
-                    gather_deepspeed3_params=self.args.ds3_gather_for_generation,
-                ) as unwrapped_model,
-                torch.no_grad(),
-                FSDP.summon_full_params(self.model_wrapped, recurse=False)
-                if self.is_fsdp_enabled
-                else nullcontext(),
-            ):
-                prompt_completion_ids = unwrapped_model.generate(
-                    input_ids=prompt_ids,
-                    attention_mask=prompt_mask,
-                    **forward_kwargs,
-                    generation_config=self.generation_config,
-                    disable_compile=True,
-                )
-            # Compute prompt length and extract completion ids
-            prompt_length = prompt_ids.size(1)
-            prompt_ids = prompt_completion_ids[:, :prompt_length]
-            completion_ids = prompt_completion_ids[:, prompt_length:]
-
-            # Mask everything after the first EOS token
-            is_eos = completion_ids == self.eos_token_id
-            eos_idx = torch.full(
-                (is_eos.size(0),), is_eos.size(1), dtype=torch.long, device=device
-            )
-            eos_idx[is_eos.any(dim=1)] = is_eos.int().argmax(dim=1)[is_eos.any(dim=1)]
-            sequence_indices = torch.arange(is_eos.size(1), device=device).expand(
-                is_eos.size(0), -1
-            )
-            completion_mask = (sequence_indices <= eos_idx.unsqueeze(1)).int()
-            prompt_ids = [p[m].tolist() for p, m in zip(prompt_ids, prompt_mask.bool())]
-            completion_ids = [
-                c[m].tolist() for c, m in zip(completion_ids, completion_mask.bool())
-            ]
-            logprobs = None  # not used in this case
-
-        return prompt_ids, completion_ids, logprobs, forward_kwargs
-
-    def _generate(self, prompts: list[str], images: Optional[list]):
-        device = self.accelerator.device
-        mode = "train" if self.model.training else "eval"
-
-        (
-            prompt_ids,
-            completion_ids,
-            logprobs,
-            forward_kwargs,
-            bob_prompt_ids,
-            bob_completion_ids,
-            bob_logprobs,
-            bob_forward_kwargs,
-            charlie_prompt_ids,
-            charlie_completion_ids,
-            charlie_logprobs,
-            charlie_forward_kwargs,
-        ) = self._generate_single_turn(prompts, images)
-
-        def log_lengths(prompt_ids, completion_ids, prefix: str):
-            prompt_lengths = torch.tensor(
-                [len(ids) for ids in prompt_ids], device=device
-            )
-            completion_lengths = torch.tensor(
-                [len(ids) for ids in completion_ids], device=device
-            )
-            agg_prompt_lengths = self.accelerator.gather(prompt_lengths)
-            agg_completion_lengths = self.accelerator.gather(completion_lengths)
-            total_prompt_tokens = agg_prompt_lengths.sum()
-            total_completion_tokens = agg_completion_lengths.sum()
-            # Log completion lengths, mean, min, max
-            self._metrics[mode][f"{prefix}/completions/mean_length"].append(
-                agg_completion_lengths.float().mean().item()
-            )
-            self._metrics[mode][f"{prefix}/completions/min_length"].append(
-                agg_completion_lengths.float().min().item()
-            )
-            self._metrics[mode][f"{prefix}/completions/max_length"].append(
-                agg_completion_lengths.float().max().item()
-            )
-            # Log the metrics
-            self.state.num_input_tokens_seen += (
-                total_prompt_tokens + total_completion_tokens
-            ).item()
-            self._metrics[mode][f"{prefix}/num_tokens"] = [
-                self.state.num_input_tokens_seen
-            ]
-
-            # Identify sequences that terminated with EOS and log their lengths
-            eos_and_pad = [self.eos_token_id, self.pad_token_id]
-            is_truncated = torch.tensor(
-                [ids[-1] not in eos_and_pad for ids in completion_ids],
-                device=device,
-            )
-            agg_is_truncated = self.accelerator.gather(is_truncated)
-            self._metrics[mode][f"{prefix}/completions/clipped_ratio"].append(
-                agg_is_truncated.float().mean().item()
-            )
-            term_completion_lengths = agg_completion_lengths[~agg_is_truncated]
-            if (
-                len(term_completion_lengths) == 0
-            ):  # edge case where no terminated sequences are found
-                term_completion_lengths = torch.zeros(1, device=device)
-            self._metrics[mode][f"{prefix}/completions/mean_terminated_length"].append(
-                term_completion_lengths.float().mean().item()
-            )
-            self._metrics[mode][f"{prefix}/completions/min_terminated_length"].append(
-                term_completion_lengths.float().min().item()
-            )
-            self._metrics[mode][f"{prefix}/completions/max_terminated_length"].append(
-                term_completion_lengths.float().max().item()
-            )
-            return total_prompt_tokens, total_completion_tokens
-
-        total_prompt_tokens, total_completion_tokens = log_lengths(
-            prompt_ids, completion_ids, "alice"
-        )
-
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        bob_total_completion_tokens = 0
-        if bob_prompt_ids is not None:
-            _, bob_total_completion_tokens = log_lengths(
-                bob_prompt_ids, bob_completion_ids, "bob"
-            )
-
-        charlie_total_completion_tokens = 0
-        if charlie_prompt_ids is not None:
-            _, charlie_total_completion_tokens = log_lengths(
-                charlie_prompt_ids, charlie_completion_ids, "charlie"
-            )
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        return (
-            prompt_ids,
-            completion_ids,
-            total_completion_tokens,
-            logprobs,
-            forward_kwargs,
-            bob_prompt_ids,
-            bob_completion_ids,
-            bob_total_completion_tokens,
-            bob_logprobs,
-            bob_forward_kwargs,
-            charlie_prompt_ids,
-            charlie_completion_ids,
-            charlie_total_completion_tokens,
-            charlie_logprobs,
-            charlie_forward_kwargs,
-        )
-
-    def _generate_bob_completions(
-        self,
-        reference_completion_ids: List[List[List[int]]],
-        reference_prompts: List[str],
-        images: Optional[list],
-    ):
-        current_active_adapter = self.model.active_adapter
-
-        self.model.set_adapter(self.bob_adapter_name)
-        try:
-            # Flatten reference_completion_ids from List[List[List[int]]] to List[List[int]]
-            # Structure: batch_size * num_references -> flat list
-            batch_size = len(reference_completion_ids)
-            num_references = len(reference_completion_ids[0]) if batch_size > 0 else 0
-            flattened_completion_ids = [
-                completion_id
-                for batch_item in reference_completion_ids
-                for completion_id in batch_item
-            ]
-
-            # Decode all at once
-            reference_solutions = self.processing_class.batch_decode(
-                flattened_completion_ids, skip_special_tokens=True
-            )
-
-            # Preprocess if needed
-            if (
-                self.preprocess_reference_solutions
-                and self.solution_preprocessor is not None
-            ):
-                reference_solutions = [
-                    self.solution_preprocessor(solution)
-                    for solution in reference_solutions
-                ]
-
-            # Reshape back to batch_size * num_references structure
-            reference_solutions = [
-                reference_solutions[i * num_references : (i + 1) * num_references]
-                for i in range(batch_size)
-            ]
-
-            prompts = self._format_bob_prompts(reference_prompts, reference_solutions)
-            # Call parent's generation with n+1
-            prompt_ids, completion_ids, logprobs, forward_kwargs = (
-                self._generate_single_turn_core(prompts, images)
-            )
-
-            # Decode
-            completions_text = self.processing_class.batch_decode(
-                completion_ids, skip_special_tokens=True
-            )
-            completions = []
-            for prompt, completion in zip(prompts, completions_text):
-                # Access bootstrap without mutating the prompt list
-                bootstrap = (
-                    prompt[-1]["content"] if prompt[-1]["role"] == "assistant" else ""
-                )
-                completions.append(
-                    [{"role": "assistant", "content": bootstrap + completion}]
-                )
-
-        finally:
-            self.model.set_adapter(current_active_adapter)
-
-        self._bob_prompts = prompts
-        self._bob_completions = completions
-        self._bob_completion_ids = completion_ids
-
-        # Bob now only generates properties, not solutions
-        self._bob_properties = self._extract_properties(completions)
-
-        return (prompt_ids, completion_ids, logprobs, forward_kwargs)
-
-    def _generate_charlie_completions(
-        self,
-        reference_completion_ids: List[List[List[int]]],
-        reference_prompts: List[str],
-        images: Optional[list],
-    ):
-        """
-        Generate Charlie's completions (alternative solutions).
-
-        Charlie receives the same inputs as Bob (problem + reference solutions)
-        but generates only alternative solutions, not properties.
-
-        Args:
-            reference_completion_ids: Reference solution token IDs
-            reference_prompts: Original problem prompts
-            images: Optional images
-
-        Returns:
-            Tuple of (prompt_ids, completion_ids, logprobs, forward_kwargs)
-        """
-        current_active_adapter = self.model.active_adapter
-
-        self.model.set_adapter(self.charlie_adapter_name)
-        try:
-            # Flatten reference_completion_ids from List[List[List[int]]] to List[List[int]]
-            # Structure: batch_size * num_references -> flat list
-            batch_size = len(reference_completion_ids)
-            num_references = len(reference_completion_ids[0]) if batch_size > 0 else 0
-            flattened_completion_ids = [
-                completion_id
-                for batch_item in reference_completion_ids
-                for completion_id in batch_item
-            ]
-
-            # Decode all at once
-            reference_solutions = self.processing_class.batch_decode(
-                flattened_completion_ids, skip_special_tokens=True
-            )
-
-            # Preprocess if needed
-            if (
-                self.preprocess_reference_solutions
-                and self.solution_preprocessor is not None
-            ):
-                reference_solutions = [
-                    self.solution_preprocessor(solution)
-                    for solution in reference_solutions
-                ]
-
-            # Reshape back to batch_size * num_references structure
-            reference_solutions = [
-                reference_solutions[i * num_references : (i + 1) * num_references]
-                for i in range(batch_size)
-            ]
-
-            prompts = self._format_charlie_prompts(
-                reference_prompts, reference_solutions
-            )
-            # Call parent's generation
-            prompt_ids, completion_ids, logprobs, forward_kwargs = (
-                self._generate_single_turn_core(prompts, images)
-            )
-
-            # Decode
-            completions_text = self.processing_class.batch_decode(
-                completion_ids, skip_special_tokens=True
-            )
-            completions = []
-            for prompt, completion in zip(prompts, completions_text):
-                # Access bootstrap without mutating the prompt list
-                bootstrap = (
-                    prompt[-1]["content"] if prompt[-1]["role"] == "assistant" else ""
-                )
-                completions.append(
-                    [{"role": "assistant", "content": bootstrap + completion}]
-                )
-
-        finally:
-            self.model.set_adapter(current_active_adapter)
-
-        self._charlie_prompts = prompts
-        self._charlie_completions = completions
-        self._charlie_completion_ids = completion_ids
-
-        # Extract solutions from Charlie's completions
-        solutions = self._extract_solutions(completions)
-        self._charlie_properties = self._extract_properties(completions)
-        self._charlie_solutions_completion_ids = [
-            self.processing_class.encode(solution, add_special_tokens=True)
-            for solution in solutions
-        ]
-        solutions = [
-            [{"role": "assistant", "content": solution}] for solution in solutions
-        ]
-        self._charlie_solutions = solutions
-
-        return (prompt_ids, completion_ids, logprobs, forward_kwargs)
-
-    def _extract_properties(self, completions: List[List[Dict[str, Any]]]) -> List[str]:
-        properties = []
-        for completion in completions:
-            try:
-                text_property = (
-                    completion[0]["content"]
-                    .split("<property>")[1]
-                    .split("</property>")[0]
-                )
-            except IndexError:
-                text_property = ""
-
-            # Apply token-length limit
-            property_ids = self.processing_class.encode(
-                text_property, add_special_tokens=False
-            )
-            if len(property_ids) > self.max_property_length:
-                text_property = self.processing_class.decode(
-                    property_ids[: self.max_property_length]
-                )
-
-            properties.append(text_property)
-        return properties
-
-    def _extract_solutions(self, completions: List[List[Dict[str, Any]]]) -> List[str]:
-        solutions = []
-        for completion in completions:
-            try:
-                solution = (
-                    completion[0]["content"]
-                    .split("<solution>")[1]
-                    .split("</solution>")[0]
-                )
-            except IndexError:
-                solution = ""
-            solutions.append(solution)
-        return solutions
-
-    def _check_properties(
-        self, solutions: List[str], properties: List[str]
-    ) -> List[bool]:
-        """
-        Check if solutions contain the specified properties using the property
-        checker adapter.
-
-        Args:
-            solutions: List of solution strings
-            properties: List of property descriptions
-
-        Returns:
-            List of boolean flags indicating if each solution has the
-            corresponding property
-        """
-        # Store current adapter
-        current_active_adapter = self.model.active_adapter
-
-        # Switch to property checker adapter
-        self.model.set_adapter(self.property_checker_adapter_name)
-
-        try:
-            if not solutions:
-                return []
-
-            prompts = []
-            for solution, property_desc in zip(solutions, properties):
-                user_message = f"PROPERTY:\n{property_desc}\n\nSOLUTION:\n{solution}"
-                prompts.append(
-                    [
-                        {
-                            "role": "system",
-                            "content": self.property_checker_prompt,
-                        },
-                        {"role": "user", "content": user_message},
-                    ]
-                )
-
-            # Process sequentially to avoid batched generation issues with left padding
-            decoded_responses = []
-            device = self.model.device
-
-            for prompt in prompts:
-                # Apply chat template to single prompt
-                prompt_text = self.processing_class.apply_chat_template(
-                    prompt,
-                    tokenize=False,
-                    add_generation_prompt=True,
-                )
-
-                # Tokenize single prompt (no padding needed)
-                model_inputs = self.processing_class(
-                    prompt_text,
-                    return_tensors="pt",
-                    padding=False,
-                    truncation=False,
-                    add_special_tokens=False,
-                ).to(device)
-
-                input_length = model_inputs["input_ids"].size(1)
-
-                # Generate for single sequence
-                with torch.inference_mode():
-                    outputs = self.model.generate(
-                        **model_inputs,
-                        max_new_tokens=16,
-                        temperature=0.1,
-                        do_sample=False,
-                    )
-
-                # Extract only the newly generated tokens
-                generated_tokens = outputs[:, input_length:]
-
-                # Decode the generated tokens
-                decoded = self.processing_class.batch_decode(
-                    generated_tokens, skip_special_tokens=True
-                )
-                decoded_responses.append(decoded[0])
-
-            # Convert to boolean: True only if property checker says "true"
-            # AND the property is not an empty string
-            # AND the solution is not an empty string
-            has_properties = [
-                (
-                    "true" in response.lower().strip()
-                    and property_desc.strip() != ""
-                    and solution.strip() != ""
-                )
-                for response, property_desc, solution in zip(
-                    decoded_responses, properties, solutions
-                )
-            ]
-
-        finally:
-            # Restore original adapter
-            self.model.set_adapter(current_active_adapter)
-
-        return has_properties
-
-    def _get_game_rewards(
-        self,
-        alice_solutions: List[str],
-        bob_solutions: List[str],
-        properties: List[str],
-        has_properties_alice: List[bool],
-        has_properties_bob: List[bool],
-        alice_scores: List[float],
-        bob_scores: List[float],
-    ) -> List[float]:
-        """
-        Compute game rewards based on the two-player game logic.
-
-        Args:
-            alice_solutions: Alice's solution strings
-            bob_solutions: Bob's solution strings
-            properties: Property descriptions from Bob
-            has_properties_alice: Whether Alice's solutions have the properties
-            has_properties_bob: Whether Bob's solutions have the properties
-            alice_scores: Test scores for Alice's solutions
-            bob_scores: Test scores for Bob's solutions
-
-        Returns:
-            List of game rewards (positive = Alice wins, negative = Bob wins)
-        """
-        score_diffs = [
-            alice_score - bob_score
-            for alice_score, bob_score in zip(alice_scores, bob_scores)
-        ]
-
-        # Compute rewards using the same logic as the original implementation
-        rewards = []
-        for i in range(len(alice_solutions)):
-            property_desc = properties[i]
-            has_property_alice = has_properties_alice[i]
-            has_property_bob = has_properties_bob[i]
-            solution_alice = alice_solutions[i]
-            solution_bob = bob_solutions[i]
-            score_diff = score_diffs[i]
-            bob_score = bob_scores[i]
-            alice_score = alice_scores[i]
-
-            # Trivial scenarios
-            if alice_score == 0.0:  # Don't score the game
-                rewards.append(0.0)
-            elif property_desc == "":  # No property
-                rewards.append(1.0)
-            elif solution_bob == "":  # No solution
-                rewards.append(1.0)
-            elif solution_alice == "":  # No solution
-                rewards.append(-1.0)
-            # Non-trivial scenarios
-            elif (
-                bob_score < self.tolerance and alice_score < self.tolerance
-            ):  # Hard problem
-                rewards.append(0.0)
-            elif not has_property_alice:  # Bad property
-                rewards.append(1.0)
-            elif has_property_bob:  # Non-removable property
-                rewards.append(1.0)
-            elif score_diff > self.tolerance:  # Alice performs better
-                rewards.append(1.0)
-            else:
-                rewards.append(-1.0)  # Bob wins
-
-        # Reward shaping: Give nuggets of reward for each won step
-        for i in range(len(rewards)):
-            alice_score = alice_scores[i]
-            bob_score = bob_scores[i]
-            alice_flagged = has_properties_alice[i]
-            bob_flagged = has_properties_bob[i]
-            bob_solution = bob_solutions[i]
-            score_val = 0.1
-
-            # Implemented a solution
-            if solution_alice == "":
-                rewards[i] = cast(float, rewards[i]) - score_val
-            else:
-                rewards[i] = cast(float, rewards[i]) + score_val
-
-            # Adversary implemented a solution
-            if bob_solution == "":
-                rewards[i] = cast(float, rewards[i]) + score_val
-            else:
-                rewards[i] = cast(float, rewards[i]) - score_val
-
-            # Implemented a solution that had a flaggable property
-            if alice_score > self.tolerance and solution_alice != "":
-                if alice_flagged:
-                    rewards[i] = cast(float, rewards[i]) - score_val
-                else:
-                    rewards[i] = cast(float, rewards[i]) + score_val
-
-            # Implemented a solution which has a removable property
-            if bob_score > self.tolerance and bob_solution != "":
-                if bob_flagged:
-                    rewards[i] = cast(float, rewards[i]) + score_val
-                else:
-                    rewards[i] = cast(float, rewards[i]) - score_val
-
-            # Implemented a solution which scored better than the other player
-            if alice_score > self.tolerance or bob_score > self.tolerance:
-                if score_diffs[i] > self.tolerance:
-                    rewards[i] = cast(float, rewards[i]) + score_val
-                else:
-                    rewards[i] = cast(float, rewards[i]) - score_val
-
-        # Apply game coefficient
-        return rewards
-
-    def _prepare_scoring_inputs(
-        self,
-        inputs,
-        prompts,
-        images,
-        prompt_ids_list,
-        completion_ids_list,
-        sampling_per_token_logps_list,
-        forward_kwargs,
-    ):
-        device = self.accelerator.device
-        mode = "train" if self.model.training else "eval"
-
-        # Convert lists of token IDs to padded tensors
-        prompt_ids = [torch.tensor(ids, device=device) for ids in prompt_ids_list]
-        prompt_mask = [torch.ones_like(ids, dtype=torch.long) for ids in prompt_ids]
-        prompt_ids = pad(
-            prompt_ids, padding_value=self.pad_token_id, padding_side="left"
-        )
-        prompt_mask = pad(prompt_mask, padding_value=0, padding_side="left")
-        completion_ids = [
-            torch.tensor(ids, device=device) for ids in completion_ids_list
-        ]
-        completion_mask = [
-            torch.ones_like(ids, dtype=torch.long) for ids in completion_ids
-        ]
-        completion_ids = pad(
-            completion_ids,
-            padding_value=self.pad_token_id,
-            padding_side="right",
-        )
-        completion_mask = pad(completion_mask, padding_value=0, padding_side="right")
-        if sampling_per_token_logps_list is not None:
-            sampling_per_token_logps = [
-                torch.tensor(logps, device=device)
-                for logps in sampling_per_token_logps_list
-            ]
-            sampling_per_token_logps = pad(
-                sampling_per_token_logps,
-                padding_value=0.0,
-                padding_side="right",
-            )
-        else:
-            sampling_per_token_logps = None
-
-        # If mask_truncated_completions is enabled, zero out truncated
-        # completions in completion_mask
-        if self.mask_truncated_completions:
-            eos_and_pad = [self.eos_token_id, self.pad_token_id]
-            is_truncated = torch.tensor(
-                [ids[-1] not in eos_and_pad for ids in completion_ids_list],
-                device=device,
-            )
-            completion_mask = completion_mask * (~is_truncated).unsqueeze(1).int()
-
-        # Concatenate prompt_mask with completion_mask for logit computation
-        prompt_completion_ids = torch.cat(
-            [prompt_ids, completion_ids], dim=1
-        )  # (B, P+C)
-        attention_mask = torch.cat([prompt_mask, completion_mask], dim=1)  # (B, P+C)
-        # If token_type_ids are used, extend them with zeros for the
-        # completion part
-        if "token_type_ids" in forward_kwargs:
-            token_type_ids = forward_kwargs["token_type_ids"]
-            forward_kwargs["token_type_ids"] = torch.cat(
-                [
-                    token_type_ids,
-                    token_type_ids.new_zeros(completion_ids.shape),
-                ],
-                dim=1,
-            )
-
-        logits_to_keep = completion_ids.size(
-            1
-        )  # we only need to compute the logits for the completion tokens
-        batch_size = (
-            self.args.per_device_train_batch_size
-            if mode == "train"
-            else self.args.per_device_eval_batch_size
-        )
-
-        num_images = (
-            [len(img_list) for img_list in images] if images is not None else None
-        )
-
-        with torch.no_grad():
-            # If the generation and optimization steps are misaligned—i.e.,
-            # if generation does not occur at the end of
-            # a full optimizer step (when gradient_accumulation_steps is not a
-            # multiple of generate_every)—then the
-            # samples may come from an earlier version of the model. In that
-            # case, we need to track old_per_token_logps
-            # for importance sampling. If the steps are aligned, importance
-            # sampling isn't necessary and we set
-            # old_per_token_logps to None.
-            # When using vLLM, we always compute old_per_token_logps for
-            # importance sampling, it was shown that the
-            # distribution mismatch between vLLM and the training model can be
-            # large and harm the training.
-            generate_every = (
-                self.args.steps_per_generation * self.num_iterations
-            )  # generation frequency
-            if self.args.gradient_accumulation_steps % generate_every != 0 or (
-                self.use_vllm and self.vllm_importance_sampling_correction
-            ):
-                old_per_token_logps, _ = self._get_per_token_logps_and_entropies(
-                    self.model,
-                    prompt_completion_ids,
-                    attention_mask,
-                    logits_to_keep,
-                    batch_size,
-                    num_images=num_images,
-                    **forward_kwargs,  # may contain pixel_values,
-                    # image_grid_thw, pixel_attention_mask and image_sizes
-                )
-            else:
-                old_per_token_logps = None
-
-            # Compute the importance sampling ratio when using vLLM, to correct
-            # for potential distribution mismatch
-            importance_sampling_ratio = None
-            if self.use_vllm and self.vllm_importance_sampling_correction:
-                importance_sampling_ratio = torch.exp(
-                    old_per_token_logps - sampling_per_token_logps
-                )
-                importance_sampling_ratio = torch.clamp(
-                    importance_sampling_ratio,
-                    max=self.vllm_importance_sampling_cap,
-                )
-
-            # Compute the per-token log probabilities for the reference model
-            if self.beta != 0.0:
-                if self.ref_model is not None:
-                    ref_per_token_logps, _ = self._get_per_token_logps_and_entropies(
-                        self.ref_model,
-                        prompt_completion_ids,
-                        attention_mask,
-                        logits_to_keep,
-                        batch_size=batch_size,
-                        num_images=num_images,
-                        **forward_kwargs,  # may contain pixel_values,
-                        # image_grid_thw, pixel_attention_mask
-                        # and image_sizes
-                    )
-                else:
-                    with self.accelerator.unwrap_model(self.model).disable_adapter():
-                        ref_per_token_logps, _ = (
-                            self._get_per_token_logps_and_entropies(
-                                self.model,
-                                prompt_completion_ids,
-                                attention_mask,
-                                logits_to_keep,
-                                batch_size=batch_size,
-                                num_images=num_images,
-                                **forward_kwargs,  # may contain pixel_values,
-                                # image_grid_thw, pixel_attention_mask
-                                # and image_sizes
-                            )
-                        )
-            else:
-                ref_per_token_logps = None
-
-        # Decode
-        prompts_text = self.processing_class.batch_decode(
-            prompt_ids, skip_special_tokens=True
-        )
-        completions_text = self.processing_class.batch_decode(
-            completion_ids, skip_special_tokens=True
-        )
-        if is_conversational(inputs[0]):
-            completions = []
-            for prompt, completion in zip(prompts, completions_text):
-                # Access bootstrap without mutating the prompt list
-                bootstrap = (
-                    prompt[-1]["content"] if prompt[-1]["role"] == "assistant" else ""
-                )
-                completions.append(
-                    [{"role": "assistant", "content": bootstrap + completion}]
-                )
-        else:
-            completions = completions_text
-
-        return (
-            prompt_ids,
-            prompt_mask,
-            completion_ids,
-            completion_mask,
-            completions,
-            prompts_text,
-            completions_text,
-            old_per_token_logps,
-            ref_per_token_logps,
-            sampling_per_token_logps,
-            importance_sampling_ratio,
-            num_images,
-        )
-
-    def _calculate_advantages(self, prompts, images, rewards_per_func):
-        device = self.accelerator.device
-        mode = "train" if self.model.training else "eval"
-        num_generations = (
-            self.num_generations
-            if mode == "train"
-            else self.num_generations_eval or self.num_generations
-        )
-        # Apply weights to each reward function's output and sum
-        rewards = (
-            rewards_per_func * self.reward_weights.to(device).unsqueeze(0)
-        ).nansum(dim=1)
-
-        # Compute grouped-wise rewards
-        mean_grouped_rewards = rewards.view(-1, num_generations).mean(dim=1)
-
-        # Normalize the rewards to compute the advantages
-        mean_grouped_rewards = mean_grouped_rewards.repeat_interleave(
-            num_generations, dim=0
-        )
-        advantages = rewards - mean_grouped_rewards
-
-        if self.scale_rewards in ["group", "none"]:
-            # If self.scale_rewards = "none", we'll still log group level std
-            std_rewards = rewards.view(-1, num_generations).std(dim=1)
-            std_rewards = std_rewards.repeat_interleave(num_generations, dim=0)
-        elif self.scale_rewards == "batch":
-            # Compute global std
-            std_rewards = rewards.std().expand_as(rewards)
-        else:
-            raise ValueError(
-                f"Invalid value for scale_rewards: {self.scale_rewards}. "
-                "Must be one of 'batch', 'group', or 'none'."
-            )
-
-        is_std_zero = torch.isclose(std_rewards, torch.zeros_like(std_rewards))
-        if self.scale_rewards != "none":
-            advantages = advantages / (std_rewards + 1e-4)
-
-        # Slice to keep only the local part of the data
-        process_slice = slice(
-            self.accelerator.process_index * len(prompts),
-            (self.accelerator.process_index + 1) * len(prompts),
-        )
-        all_process_advantages = (
-            advantages.clone()
-        )  # keep the aggregated advantages for logging
-        advantages = advantages[process_slice]
-
-        return (
-            advantages,
-            all_process_advantages,
-            std_rewards,
-            is_std_zero,
-            mean_grouped_rewards,
-        )
-
-    def _generate_and_score_completions(
-        self, inputs: list[dict[str, Union[torch.Tensor, Any]]]
-    ) -> dict[str, Union[torch.Tensor, Any]]:
-        device = self.accelerator.device
-        mode = "train" if self.model.training else "eval"
-
-        prompts = [x["prompt"] for x in inputs]
-
-        if "images" in inputs[0]:
-            images = [example.get("images") for example in inputs]
-        elif "image" in inputs[0]:
-            images = [
-                [example.get("image")] if example.get("image") is not None else None
-                for example in inputs
-            ]
-        else:
-            images = None
-
-        (
-            prompt_ids_list,
-            completion_ids_list,
-            num_items_in_batch,
-            sampling_per_token_logps_list,
-            forward_kwargs,
-            # Assume for evals these are None
-            bob_prompt_ids_list,
-            bob_completion_ids_list,
-            bob_num_items_in_batch,
-            bob_sampling_per_token_logps_list,
-            bob_forward_kwargs,
-            charlie_prompt_ids_list,
-            charlie_completion_ids_list,
-            charlie_num_items_in_batch,
-            charlie_sampling_per_token_logps_list,
-            charlie_forward_kwargs,
-        ) = self._generate(prompts, images)
-
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            # Filter out the reference completions from the training completions
-            # Now we have 2 * num_references (one set for Bob, one set for Charlie)
-            total_per_batch = self.num_generations + 2 * self.num_references
-            train_mask = [
-                i % total_per_batch < self.num_generations for i in range(len(prompts))
-            ]
-            inputs = [inputs[i] for i in range(len(inputs)) if train_mask[i]]
-            prompts = [prompts[i] for i in range(len(prompts)) if train_mask[i]]
-            prompt_ids_list = [
-                prompt_ids_list[i] for i in range(len(prompt_ids_list)) if train_mask[i]
-            ]
-            completion_ids_list = [
-                completion_ids_list[i]
-                for i in range(len(completion_ids_list))
-                if train_mask[i]
-            ]
-            if sampling_per_token_logps_list is not None:
-                sampling_per_token_logps_list = [
-                    sampling_per_token_logps_list[i]
-                    for i in range(len(sampling_per_token_logps_list))
-                    if train_mask[i]
-                ]
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        (
-            prompt_ids,
-            prompt_mask,
-            completion_ids,
-            completion_mask,
-            completions,
-            prompts_text,
-            completions_text,
-            old_per_token_logps,
-            ref_per_token_logps,
-            sampling_per_token_logps,
-            importance_sampling_ratio,
-            num_images,
-        ) = self._prepare_scoring_inputs(
-            inputs,
-            prompts,
-            images,
-            prompt_ids_list,
-            completion_ids_list,
-            sampling_per_token_logps_list,
-            forward_kwargs,
-        )
-
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            bob_prompts = self._bob_prompts
-            bob_inputs = [{"prompt": prompt} for prompt in bob_prompts]
-            # Set requires grad to True for Bob's adapter
-            original_adapter = self.model.active_adapter
-            self.model.set_adapter(self.bob_adapter_name)
-            try:
-                (
-                    bob_prompt_ids,
-                    bob_prompt_mask,
-                    bob_completion_ids,
-                    bob_completion_mask,
-                    bob_completions,
-                    bob_prompts_text,
-                    bob_completions_text,
-                    bob_old_per_token_logps,
-                    bob_ref_per_token_logps,
-                    bob_sampling_per_token_logps,
-                    bob_importance_sampling_ratio,
-                    bob_num_images,
-                ) = self._prepare_scoring_inputs(
-                    bob_inputs,
-                    bob_prompts,
-                    images,
-                    bob_prompt_ids_list,
-                    bob_completion_ids_list,
-                    bob_sampling_per_token_logps_list,
-                    bob_forward_kwargs,
-                )
-            finally:
-                # Restore the original adapter
-                self.model.set_adapter(original_adapter)
-
-            # Prepare Charlie's inputs similarly
-            charlie_prompts = self._charlie_prompts
-            charlie_inputs = [{"prompt": prompt} for prompt in charlie_prompts]
-            self.model.set_adapter(self.charlie_adapter_name)
-            try:
-                (
-                    charlie_prompt_ids,
-                    charlie_prompt_mask,
-                    charlie_completion_ids,
-                    charlie_completion_mask,
-                    charlie_completions,
-                    charlie_prompts_text,
-                    charlie_completions_text,
-                    charlie_old_per_token_logps,
-                    charlie_ref_per_token_logps,
-                    charlie_sampling_per_token_logps,
-                    charlie_importance_sampling_ratio,
-                    charlie_num_images,
-                ) = self._prepare_scoring_inputs(
-                    charlie_inputs,
-                    charlie_prompts,
-                    images,
-                    charlie_prompt_ids_list,
-                    charlie_completion_ids_list,
-                    charlie_sampling_per_token_logps_list,
-                    charlie_forward_kwargs,
-                )
-            finally:
-                # Restore the original adapter
-                self.model.set_adapter(original_adapter)
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        # Calculate rewards for each reward function. rewards_per_func aggregates rewards across all processes. This is
-        # important because rewards will be normalized per group, and completions are distributed. We will later slice
-        # rewards_per_func to extract each process's subset.
-        rewards_per_func = self._calculate_rewards(
-            inputs, prompts, completions, completion_ids_list
-        )
-
-        (
-            advantages,
-            all_process_advantages,
-            std_rewards,
-            is_std_zero,
-            mean_grouped_rewards,
-        ) = self._calculate_advantages(prompts, images, rewards_per_func)
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        # We compute bob's and charlie's advantages from negated game rewards (only during training)
-        # Bob and Charlie are on the same team, so they both get negative game rewards
-        # All other rewards are zeroed out for Bob and Charlie
-        if self.model.training:
-            bob_rewards_per_func = rewards_per_func.clone()
-            bob_rewards_per_func[:, :-1] = 0
-            bob_rewards_per_func[:, -1] = -bob_rewards_per_func[:, -1]
-            bob_advantages, *_ = self._calculate_advantages(
-                prompts, images, bob_rewards_per_func
-            )
-
-            # Charlie uses the same advantages as Bob (same team)
-            charlie_advantages = bob_advantages.clone()
-        ############### END GRPOGAMETRAINER CHANGE ###############
-
-        # Calculate mean reward per function, but only for samples where the function was applied (non-NaN values)
-        # Only iterate over the actual columns present in rewards_per_func (may exclude game_reward during eval)
-        num_reward_cols = rewards_per_func.shape[1]
-        for i, reward_func_name in enumerate(self.reward_func_names[:num_reward_cols]):
-            mean_rewards = torch.nanmean(rewards_per_func[:, i]).item()
-            self._metrics[mode][f"rewards/{reward_func_name}/mean"].append(mean_rewards)
-            std_func_rewards = nanstd(rewards_per_func[:, i]).item()
-            self._metrics[mode][f"rewards/{reward_func_name}/std"].append(
-                std_func_rewards
-            )
-        self._metrics[mode]["reward"].append(mean_grouped_rewards.mean().item())
-        self._metrics[mode]["reward_std"].append(std_rewards.mean().item())
-        self._metrics[mode]["frac_reward_zero_std"].append(
-            is_std_zero.float().mean().item()
-        )
-
-        # Log prompt and completion texts
-        self._logs["prompt"].extend(gather_object(prompts_text))
-        self._logs["completion"].extend(gather_object(completions_text))
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            self._logs["bob_prompt"].extend(gather_object(bob_prompts_text))
-            self._logs["bob_completion"].extend(gather_object(bob_completions_text))
-            self._logs["charlie_prompt"].extend(gather_object(charlie_prompts_text))
-            self._logs["charlie_completion"].extend(
-                gather_object(charlie_completions_text)
-            )
-        ############### END GRPOGAMETRAINER CHANGE ###############
-        # Only log rewards for columns that actually exist in rewards_per_func (may exclude game_reward during eval)
-        num_reward_cols = rewards_per_func.shape[1]
-        for i, name in enumerate(self.reward_func_names[:num_reward_cols]):
-            self._logs["rewards"][name].extend(rewards_per_func[:, i].tolist())
-        self._logs["advantages"].extend(all_process_advantages.tolist())
-
-        if images is not None:
-            self._logs["images"].extend(gather_object(images))
-
-        if self.use_vllm and self.vllm_importance_sampling_correction:
-            delta = torch.abs(old_per_token_logps - sampling_per_token_logps)
-            delta = delta[completion_mask.bool()]
-            mean_delta = (
-                torch.mean(delta)
-                if delta.numel() > 0
-                else torch.tensor(0.0, device=device)
-            )
-            max_delta = (
-                torch.max(delta)
-                if delta.numel() > 0
-                else torch.tensor(0.0, device=device)
-            )
-            self._metrics[mode]["sampling/sampling_logp_difference/mean"].append(
-                self.accelerator.gather(mean_delta).mean().item()
-            )
-            self._metrics[mode]["sampling/sampling_logp_difference/max"].append(
-                self.accelerator.gather(max_delta).max().item()
-            )
-
-            flat_is_ratio = importance_sampling_ratio[completion_mask.bool()]
-            min_importance_sampling_ratio = (
-                torch.min(flat_is_ratio)
-                if flat_is_ratio.numel() > 0
-                else torch.tensor(0.0, device=device)
-            )
-            mean_importance_sampling_ratio = (
-                torch.mean(flat_is_ratio)
-                if flat_is_ratio.numel() > 0
-                else torch.tensor(0.0, device=device)
-            )
-            max_importance_sampling_ratio = (
-                torch.max(flat_is_ratio)
-                if flat_is_ratio.numel() > 0
-                else torch.tensor(0.0, device=device)
-            )
-            self._metrics[mode]["sampling/importance_sampling_ratio/min"].append(
-                nanmin(self.accelerator.gather(min_importance_sampling_ratio)).item()
-            )
-            self._metrics[mode]["sampling/importance_sampling_ratio/mean"].append(
-                self.accelerator.gather(mean_importance_sampling_ratio).nanmean().item()
-            )
-            self._metrics[mode]["sampling/importance_sampling_ratio/max"].append(
-                nanmax(self.accelerator.gather(max_importance_sampling_ratio)).item()
-            )
-
-        output = {
-            "prompt_ids": prompt_ids,
-            "prompt_mask": prompt_mask,
-            "completion_ids": completion_ids,
-            "completion_mask": completion_mask,
-            "advantages": advantages,
-            "num_items_in_batch": num_items_in_batch,
-        }
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            output.update(
-                {
-                    "bob_prompt_ids": bob_prompt_ids,
-                    "bob_prompt_mask": bob_prompt_mask,
-                    "bob_completion_ids": bob_completion_ids,
-                    "bob_completion_mask": bob_completion_mask,
-                    "bob_advantages": bob_advantages,
-                    "bob_num_items_in_batch": bob_num_items_in_batch,
-                    "charlie_prompt_ids": charlie_prompt_ids,
-                    "charlie_prompt_mask": charlie_prompt_mask,
-                    "charlie_completion_ids": charlie_completion_ids,
-                    "charlie_completion_mask": charlie_completion_mask,
-                    "charlie_advantages": charlie_advantages,
-                    "charlie_num_items_in_batch": charlie_num_items_in_batch,
-                }
-            )
-        ############### END GRPOGAMETRAINER CHANGE ###############
-        if old_per_token_logps is not None:
-            output["old_per_token_logps"] = old_per_token_logps
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            if bob_old_per_token_logps is not None:
-                output["bob_old_per_token_logps"] = bob_old_per_token_logps
-            if charlie_old_per_token_logps is not None:
-                output["charlie_old_per_token_logps"] = charlie_old_per_token_logps
-        ############### END GRPOGAMETRAINER CHANGE ###############
-        if self.use_vllm and self.vllm_importance_sampling_correction:
-            output["importance_sampling_ratio"] = importance_sampling_ratio
-            ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-            if self.model.training:
-                output["bob_importance_sampling_ratio"] = bob_importance_sampling_ratio
-                output["charlie_importance_sampling_ratio"] = (
-                    charlie_importance_sampling_ratio
-                )
-            ############### END GRPOGAMETRAINER CHANGE ###############
-        if ref_per_token_logps is not None:
-            output["ref_per_token_logps"] = ref_per_token_logps
-        ############### BEGIN GRPOGAMETRAINER CHANGE ###############
-        if self.model.training:
-            if bob_ref_per_token_logps is not None:
-                output["bob_ref_per_token_logps"] = bob_ref_per_token_logps
-            if charlie_ref_per_token_logps is not None:
-                output["charlie_ref_per_token_logps"] = charlie_ref_per_token_logps
-        ############### END GRPOGAMETRAINER CHANGE ###############
-        if "pixel_values" in forward_kwargs:
-            output["pixel_values"] = forward_kwargs["pixel_values"]
-        if "image_grid_thw" in forward_kwargs:
-            output["image_grid_thw"] = forward_kwargs["image_grid_thw"]
-        if "pixel_attention_mask" in forward_kwargs:
-            output["pixel_attention_mask"] = forward_kwargs["pixel_attention_mask"]
-        if "image_sizes" in forward_kwargs:
-            output["image_sizes"] = forward_kwargs["image_sizes"]
-        if "token_type_ids" in forward_kwargs:
-            output["token_type_ids"] = forward_kwargs["token_type_ids"]
-        if images is not None:
-            output["num_images"] = num_images
-        return output
-
-    def compute_liger_loss(self, unwrapped_model, inputs):
-        # Compute the per-token log probabilities for the model
-        prompt_ids, prompt_mask = inputs["prompt_ids"], inputs["prompt_mask"]
-        completion_ids, completion_mask = (
-            inputs["completion_ids"],
-            inputs["completion_mask"],
-        )
-        input_ids = torch.cat([prompt_ids, completion_ids], dim=1)
-        attention_mask = torch.cat([prompt_mask, completion_mask], dim=1)
-        logits_to_keep = completion_ids.size(
-            1
-        )  # we only need to compute the logits for the completion tokens
-
-        # Get the last hidden state of the model
-        last_hidden_state = self._get_last_hidden_state(
-            unwrapped_model,
-            input_ids,
-            attention_mask,
-            logits_to_keep,
-            inputs.get("pixel_values"),
-            inputs.get("image_grid_thw"),
-            inputs.get("pixel_attention_mask"),
-            inputs.get("image_sizes"),
-        )
-
-        # compute loss and metrics using liger grpo loss
-        loss, metrics = self.liger_grpo_loss(
-            _input=last_hidden_state,
-            lin_weight=unwrapped_model.lm_head.weight,
-            selected_token_ids=completion_ids,
-            attention_mask=completion_mask,
-            advantages=inputs["advantages"],
-            bias=unwrapped_model.lm_head.bias,
-            old_per_token_logps=inputs.get("old_per_token_logps"),
-            ref_per_token_logps=inputs.get("ref_per_token_logps"),
-        )
-        # Extract metrics from the liger_grpo_loss output
-        # KL divergence is the first metric when beta is non-zero
-        mean_kl = metrics[0] if self.beta != 0.0 else None
-        clip_ratio = metrics[-1]
-
-        mode = "train" if self.model.training else "eval"
-        if self.beta != 0.0:
-            self._metrics[mode]["kl"].append(
-                self.accelerator.gather(mean_kl).mean().item()
-            )
-        self._metrics[mode]["clip_ratio"].append(
-            self.accelerator.gather(clip_ratio).mean().item()
-        )
-        return loss / self.current_gradient_accumulation_steps
-
-    @profiling_decorator
-    def compute_loss(
-        self, model, inputs, return_outputs=False, num_items_in_batch=None
-    ):
-        """
-        Override compute_loss to train Alice, Bob, and Charlie simultaneously.
-
-        This method:
-        1. Computes Alice's loss with existing game reward advantages
-        2. Switches to Bob's adapter and computes Bob's loss (property identification)
-        3. Switches to Charlie's adapter and computes Charlie's loss (alternative solutions)
-        4. Bob and Charlie both use negative game rewards (same team)
-        5. Returns all three losses
-
-        When train_adversary=False, only Alice's loss is computed and Bob/Charlie losses are None.
-
-        Args:
-            model: The model being trained
-            inputs: Input dictionary
-            return_outputs: Whether to return outputs
-            (not supported for combined loss)
-            num_items_in_batch: Number of items in the batch
-        Returns:
-            Tuple of (alice_loss, bob_loss, charlie_loss) where bob_loss and charlie_loss
-            are None if train_adversary=False
-        """
-        # During evaluation, skip all game mechanisms and use parent's implementation
-        if not self.model.training:
-            return self.compute_loss_core(
-                model, inputs, return_outputs, num_items_in_batch
-            )
-
-        # Store the original active adapter
-        original_adapter = self.model.active_adapter
-
-        # Compute Alice's loss with the existing advantages from inputs
-        # (These advantages were computed from game rewards
-        # in _calculate_rewards)
-        alice_loss = self.compute_loss_core(model, inputs, return_outputs=False)
-
-        # Skip Bob and Charlie loss computation if train_adversary is disabled
-        if not self.train_adversary:
-            return alice_loss, None, None
-
-        # Switch to Bob's adapter to train property identification
-        self.model.set_adapter(self.bob_adapter_name)
-        try:
-            bob_inputs = inputs.copy()
-            # Construct Bob's inputs
-            for key, value in inputs.items():
-                if key.startswith("bob_"):
-                    bob_inputs[key.replace("bob_", "")] = value
-
-            # Compute Bob's loss
-            bob_loss = self.compute_loss_core(model, bob_inputs, return_outputs=False)
-        finally:
-            # Restore the original adapter
-            self.model.set_adapter(original_adapter)
-
-        # Switch to Charlie's adapter to train alternative solution generation
-        self.model.set_adapter(self.charlie_adapter_name)
-        try:
-            charlie_inputs = inputs.copy()
-            # Construct Charlie's inputs (use Bob's advantages since they're on the same team)
-            for key, value in inputs.items():
-                if key.startswith("charlie_"):
-                    charlie_inputs[key.replace("charlie_", "")] = value
-
-            # Compute Charlie's loss
-            charlie_loss = self.compute_loss_core(
-                model, charlie_inputs, return_outputs=False
-            )
-        finally:
-            # Restore the original adapter
-            self.model.set_adapter(original_adapter)
-
-        return alice_loss, bob_loss, charlie_loss
-
-    def compute_loss_core(
-        self, model, inputs, return_outputs=False, num_items_in_batch=None
-    ):
-        if return_outputs:
-            raise ValueError("The GRPOTrainer does not support returning outputs")
-        if self.use_liger_loss:
-            # Compute the loss using the liger grpo loss
-            unwrapped_model = self.accelerator.unwrap_model(model)
-            return self._forward_redirection(
-                model, unwrapped_model, self.compute_liger_loss, unwrapped_model, inputs
-            )
-        else:
-            return self._compute_loss(model, inputs)
-
-    def _compute_loss(self, model, inputs):
-        # Compute the per-token log probabilities for the model
-        prompt_ids, prompt_mask = inputs["prompt_ids"], inputs["prompt_mask"]
-        completion_ids, completion_mask = (
-            inputs["completion_ids"],
-            inputs["completion_mask"],
-        )
-        input_ids = torch.cat([prompt_ids, completion_ids], dim=1)
-        attention_mask = torch.cat([prompt_mask, completion_mask], dim=1)
-        logits_to_keep = completion_ids.size(
-            1
-        )  # we only need to compute the logits for the completion tokens
-
-        # Compute the per_token_logps and the entropy at each position in the completion
-        per_token_logps, entropies = self._get_per_token_logps_and_entropies(
-            model,
-            input_ids,
-            attention_mask,
-            logits_to_keep,
-            compute_entropy=True,
-            pixel_values=inputs.get("pixel_values"),
-            image_grid_thw=inputs.get("image_grid_thw"),
-            num_images=inputs.get("num_images"),
-            pixel_attention_mask=inputs.get("pixel_attention_mask"),
-            image_sizes=inputs.get("image_sizes"),
-            token_type_ids=inputs.get("token_type_ids"),
-        )
-
-        if self.top_entropy_quantile < 1.0:
-            entropy_mask = self.get_high_entropy_mask(
-                entropies, completion_mask, 1 - self.top_entropy_quantile
-            )
-        else:
-            entropy_mask = None
-
-        # Compute the KL divergence between the model and the reference model
-        if self.beta != 0.0:
-            ref_per_token_logps = inputs["ref_per_token_logps"]
-            per_token_kl = (
-                torch.exp(ref_per_token_logps - per_token_logps)
-                - (ref_per_token_logps - per_token_logps)
-                - 1
-            )
-
-        # Compute the loss
-        advantages = inputs["advantages"]
-        # When num_iterations == 1 and steps_per_generation <= gradient_accumulation_steps,
-        # old_per_token_logps == per_token_logps. In this case we can skip its computation
-        # (see _generate_and_score_completions) and instead use per_token_logps.detach().
-        # The exception is when using vLLM, where we always compute old_per_token_logps
-        # for importance sampling
-        old_per_token_logps = inputs.get("old_per_token_logps")
-        old_per_token_logps = (
-            per_token_logps.detach()
-            if old_per_token_logps is None
-            else old_per_token_logps
-        )
-
-        log_ratio = per_token_logps - old_per_token_logps
-        if self.importance_sampling_level == "token":
-            log_importance_weights = log_ratio
-        elif self.importance_sampling_level == "sequence":
-            log_importance_weights = (log_ratio * completion_mask).sum(
-                -1
-            ) / completion_mask.sum(-1).clamp(min=1.0)
-            log_importance_weights = log_importance_weights.unsqueeze(-1)
-        else:
-            raise ValueError(
-                f"Unknown importance sampling level: {self.importance_sampling_level}. Possible values are 'token' "
-                "and 'sequence'."
-            )
-        # From here, log_importance_weights (and all subsequent tensors, coef_1, coef_2, etc.) shape depends on
-        # importance_sampling_level: "token" level: (B, T); "sequence" level: (B, 1)
-
-        coef_1 = torch.exp(log_importance_weights)
-        coef_2 = torch.clamp(coef_1, 1 - self.epsilon_low, 1 + self.epsilon_high)
-
-        # Two-sided clipping
-        if self.args.delta is not None:
-            coef_1 = torch.clamp(coef_1, max=self.args.delta)
-
-        per_token_loss1 = coef_1 * advantages.unsqueeze(1)
-        per_token_loss2 = coef_2 * advantages.unsqueeze(1)
-        per_token_loss = -torch.min(per_token_loss1, per_token_loss2)
-        if entropy_mask is not None:
-            per_token_loss = per_token_loss * entropy_mask
-
-        if self.use_vllm and self.vllm_importance_sampling_correction:
-            per_token_loss = per_token_loss * inputs["importance_sampling_ratio"]
-
-        if self.beta != 0.0:
-            per_token_loss = per_token_loss + self.beta * per_token_kl
-
-        if self.loss_type == "grpo":
-            loss = (
-                (per_token_loss * completion_mask).sum(-1)
-                / completion_mask.sum(-1).clamp(min=1.0)
-            ).mean()
-            loss = loss / self.current_gradient_accumulation_steps
-        elif self.loss_type == "bnpo":
-            loss = (
-                per_token_loss * completion_mask
-            ).sum() / completion_mask.sum().clamp(min=1.0)
-            loss = loss / self.current_gradient_accumulation_steps
-        elif self.loss_type == "dr_grpo":
-            loss = (per_token_loss * completion_mask).sum() / (
-                per_token_loss.size(0) * self.max_completion_length
-            )
-            loss = loss / self.current_gradient_accumulation_steps
-        elif self.loss_type == "dapo":
-            normalizer = inputs["num_items_in_batch"] / self.accelerator.num_processes
-            loss = (per_token_loss * completion_mask).sum() / normalizer
-        else:
-            raise ValueError(f"Unknown loss type: {self.loss_type}")
-
-        # Log the metrics
-        mode = "train" if self.model.training else "eval"
-
-        completion_token_count = completion_mask.sum().clamp(min=1.0)
-
-        def masked_batch_mean(x):
-            if x.shape[1] == 1:  # when importance_sampling_level == "sequence"
-                return x.mean()
-            else:
-                return (x * completion_mask).sum() / completion_token_count
-
-        if self.beta != 0.0:
-            mean_kl = masked_batch_mean(per_token_kl)
-            self._metrics[mode]["kl"].append(
-                self.accelerator.gather(mean_kl).nanmean().item()
-            )
-
-        mean_entropy = masked_batch_mean(entropies)
-        self._metrics[mode]["entropy"].append(
-            self.accelerator.gather(mean_entropy).nanmean().item()
-        )
-
-        # Compute the clipped probability ratios
-        is_low_clipped = (coef_1 < 1 - self.epsilon_low) & (advantages.unsqueeze(1) < 0)
-        is_high_clipped = (coef_1 > 1 + self.epsilon_high) & (
-            advantages.unsqueeze(1) > 0
-        )
-        is_region_clipped = is_low_clipped | is_high_clipped
-
-        low_clip = masked_batch_mean(is_low_clipped.float())
-        high_clip = masked_batch_mean(is_high_clipped.float())
-        clip_ratio = masked_batch_mean(is_region_clipped.float())
-
-        gathered_low_clip = self.accelerator.gather(low_clip)
-        self._metrics[mode]["clip_ratio/low_mean"].append(
-            gathered_low_clip.nanmean().item()
-        )
-        self._metrics[mode]["clip_ratio/low_min"].append(
-            nanmin(gathered_low_clip).item()
-        )
-        gathered_high_clip = self.accelerator.gather(high_clip)
-        self._metrics[mode]["clip_ratio/high_mean"].append(
-            gathered_high_clip.nanmean().item()
-        )
-        self._metrics[mode]["clip_ratio/high_max"].append(
-            nanmax(gathered_high_clip).item()
-        )
-        gathered_clip_ratio = self.accelerator.gather(clip_ratio)
-        self._metrics[mode]["clip_ratio/region_mean"].append(
-            gathered_clip_ratio.nanmean().item()
-        )
-        return loss
-
-    def training_step(
-        self,
-        model: nn.Module,
-        inputs: dict[str, Union[torch.Tensor, Any]],
-        num_items_in_batch: Optional[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        """
-        Perform a training step on a batch of inputs.
-
-        Subclass and override to inject custom behavior.
-
-        Args:
-            model (`nn.Module`):
-                The model to train.
-            inputs (`dict[str, Union[torch.Tensor, Any]]`):
-                The inputs and targets of the model.
-
-                The dictionary will be unpacked before being fed to the model.
-                Most models expect the targets under the
-                argument `labels`. Check your model's documentation
-                for all accepted arguments.
-
-        Return:
-            `torch.Tensor`: The tensor with training loss on this batch.
-        """
-        # Prepare buffers for context parallelism
-
-        cp_context, inputs = self._prepare_context_parallel_inputs(model, inputs)
-
-        # Context manager is no-op if CP isn't enabled
-        with cp_context():
-            model.train()
-            if hasattr(self.optimizer, "train") and callable(self.optimizer.train):
-                self.optimizer.train()
-
-            inputs = self._prepare_inputs(inputs)
-            if is_sagemaker_mp_enabled():
-                loss_mb = smp_forward_backward(
-                    model, inputs, self.args.gradient_accumulation_steps
-                )
-                return loss_mb.reduce_mean().detach().to(self.args.device)
-
-            with self.compute_loss_context_manager():
-                alice_loss, bob_loss, charlie_loss = self.compute_loss(
-                    model, inputs, num_items_in_batch=num_items_in_batch
-                )
-
-            del inputs
-            if (
-                self.args.torch_empty_cache_steps is not None
-                and self.state.global_step % self.args.torch_empty_cache_steps == 0
-            ):
-                if is_torch_xpu_available():
-                    torch.xpu.empty_cache()
-                elif is_torch_mlu_available():
-                    torch.mlu.empty_cache()
-                elif is_torch_musa_available():
-                    torch.musa.empty_cache()
-                elif is_torch_npu_available():
-                    torch.npu.empty_cache()
-                elif is_torch_mps_available():
-                    torch.mps.empty_cache()
-                elif is_torch_hpu_available():
-                    logger.warning(
-                        "`torch_empty_cache_steps` is set but HPU "
-                        "device/backend does not support empty_cache()."
-                    )
-                else:
-                    torch.cuda.empty_cache()
-
-            kwargs = {}
-
-            # For LOMO optimizers you need to explicitly use the learning rate
-            if self.args.optim in [
-                OptimizerNames.LOMO,
-                OptimizerNames.ADALOMO,
-            ]:
-                kwargs["learning_rate"] = self._get_learning_rate()
-
-            def loss_backward(loss):
-                if self.args.n_gpu > 1:
-                    loss = (
-                        loss.mean()
-                    )  # mean() to average on multi-gpu parallel training
-
-                if self.use_apex:
-                    from apex import amp
-
-                    with amp.scale_loss(loss, self.optimizer) as scaled_loss:
-                        scaled_loss.backward()
-                else:
-                    # Finally we need to normalize the loss for reporting
-                    # if GA loss bug is not fixed during compute loss
-                    if (
-                        not self.model_accepts_loss_kwargs or num_items_in_batch is None
-                    ) and self.compute_loss_func is None:
-                        # If the model does not accept loss kwargs, we need to
-                        # normalize the loss by the number of gradient
-                        # accumulation steps
-                        loss = loss / self.current_gradient_accumulation_steps
-
-                    # Turning off loss scaling w.r.t. gradient accumulation
-                    # when DeepSpeed is enabled
-                    # https://github.com/huggingface/transformers/pull/35808
-                    if self.accelerator.distributed_type == DistributedType.DEEPSPEED:
-                        kwargs["scale_wrt_gas"] = False
-
-                    self.accelerator.backward(loss, **kwargs)
-
-                return loss.detach()
-
-            original_adapter = self.model.active_adapter
-            self.model.set_adapter(original_adapter)
-            alice_loss = loss_backward(alice_loss)
-
-            # Only compute backward passes for Bob and Charlie if train_adversary is enabled
-            if self.train_adversary and bob_loss is not None:
-                self.model.set_adapter(self.bob_adapter_name)
-                bob_loss = loss_backward(bob_loss)
-            if self.train_adversary and charlie_loss is not None:
-                self.model.set_adapter(self.charlie_adapter_name)
-                charlie_loss = loss_backward(charlie_loss)
-
-            self.model.set_adapter(original_adapter)
-
-            return alice_loss
-
-    def prediction_step(
-        self,
-        model,
-        inputs,
-        prediction_loss_only,
-        ignore_keys: Optional[list[str]] = None,
-    ):
-        inputs = self._prepare_inputs(inputs)
-        with torch.no_grad():
-            with self.compute_loss_context_manager():
-                loss = self.compute_loss(model, inputs)
-            loss = loss.mean().detach()
-        return loss, None, None
-
-    def log(self, logs: dict[str, float], start_time: Optional[float] = None) -> None:
-        mode = "train" if self.model.training else "eval"
-        metrics = {
-            key: sum(val) / len(val) for key, val in self._metrics[mode].items()
-        }  # average the metrics
-
-        # This method can be called both in training and evaluation. When called in evaluation, the keys in `logs`
-        # start with "eval_". We need to add the prefix "eval_" to the keys in `metrics` to match the format.
-        if mode == "eval":
-            metrics = {f"eval_{key}": val for key, val in metrics.items()}
-
-        logs = {**logs, **metrics}
-        super().log(logs, start_time)
-        self._metrics[mode].clear()
-
-        if self.accelerator.is_main_process and self.log_completions:
-            if is_rich_available():
-                print_prompt_completions_sample(
-                    self._logs["prompt"],
-                    self._logs["completion"],
-                    self._logs["rewards"],
-                    self._logs["advantages"],
-                    self.state.global_step,
-                    self.num_completions_to_print,
-                )
-
-            if (
-                self.args.report_to
-                and "wandb" in self.args.report_to
-                and wandb.run is not None
-            ):
-                import pandas as pd
-
-                table = {
-                    "step": [str(self.state.global_step)] * len(self._logs["prompt"]),
-                    "prompt": self._logs["prompt"],
-                    "completion": self._logs["completion"],
-                    **self._logs["rewards"],
-                    "advantage": self._logs["advantages"],
-                }
-
-                if self._logs["images"]:
-                    table["images"] = []
-                    for image_list in self._logs["images"]:
-                        # Convert images to wandb Image objects for proper visualization
-                        table["images"].append(
-                            [wandb.Image(image) for image in image_list]
-                        )
-
-                df = pd.DataFrame(table)
-                if self.wandb_log_unique_prompts:
-                    df = df.drop_duplicates(subset=["prompt"])
-                wandb.log({"completions": wandb.Table(dataframe=df)})
-
-    def create_optimizer(self):
-        """
-        Override to ensure all game adapter parameters have correct requires_grad flags before optimizer creation.
-        """
-        # Ensure game adapters have correct requires_grad before optimizer sees them
-        game_adapters = [
-            self.property_checker_adapter_name,
-            self.bob_adapter_name,
-            self.charlie_adapter_name,
-        ]
-
-        for adapter_name in game_adapters:
-            if (
-                hasattr(self.model, "peft_config")
-                and adapter_name in self.model.peft_config
-            ):
-                is_trainable = (
-                    adapter_name in [self.bob_adapter_name, self.charlie_adapter_name]
-                    and self.train_adversary
-                )
-                for name, param in self.model.named_parameters():
-                    if adapter_name in name:
-                        param.requires_grad = is_trainable
-
-        # Log trainable param count before creating optimizer
-        trainable_params = [p for p in self.model.parameters() if p.requires_grad]
-        logger.info(
-            f"Creating optimizer with {len(trainable_params)} trainable parameters"
-        )
-
-        return super().create_optimizer()
-
-    # Ensure the model card is saved along with the checkpoint
-    def _save_checkpoint(self, model, trial):
-        if self.args.hub_model_id is None:
-            model_name = Path(self.args.output_dir).name
-        else:
-            model_name = self.args.hub_model_id.split("/")[-1]
-        self.create_model_card(model_name=model_name)
-        super()._save_checkpoint(model, trial)
